starcatalogs.jl

Missing docs:

SOFA.hfk5z

SOFA.h2fk5

SOFA.fk5hz

SOFA.fk5hip

SOFA.fk54z

SOFA.fk52h

SOFA.fk524

SOFA.fk45z

Astrometry.SOFA.fk425 — Function

fk425(ra::AbstractFloat, ded::AbstractFloat, δra::AbstractFloat, δdec::AbstractFloat, plx::AbstractFloat,
      rv::AbstractFloat)

Convert B1950.0 FK4 star catalog data to J2000.0 FK5.

This function converts a star's catalog data from the old FK4 (Bessel-Newcomb) system to the later IAU 1976 FK5 (Fricke) system.

Input (all B1950.0, FK4)

ra – B1950.0 RA (rad)

dec – B1950.0 Dec (rad)

δra – B1950.0 RA proper moation (rad/trop-year)

δdec – B1950.0 Dec proper motions (rad/trop-year)

plx – parallax (arcsec)

rv – radial velocity (km/s, +ve = moving away)

Output (all J2000.0, FK5)

ra – J2000.0 RA (rad)

dec – J2000.0 Dec (rad)

δra – J2000.0 RA proper motion (rad/Jul-year)

δdec – J2000.0 Dec proper motions (rad/Jul-year)

plx – parallax (arcsec)

rv – radial velocity (km/s, +ve = moving away)

Note

The proper motions in RA are dRA/dt rather than cos(Dec)*dRA/dt, and are per year rather than per century.

The conversion is somewhat complicated, for several reasons:

. Change of standard epoch from B1950.0 to J2000.0.

. An intermediate transition date of 1984 January 1.0 TT.

. A change of precession model.

. Change of time unit for proper motion (tropical to Julian).

. FK4 positions include the E-terms of aberration, to simplify the hand computation of annual aberration. FK5 positions assume a rigorous aberration computation based on the Earth's barycentric velocity.

. The E-terms also affect proper motions, and in particular cause objects at large distances to exhibit fictitious proper motions.

The algorithm is based on Smith et al. (1989) and Yallop et al. (1989), which presented a matrix method due to Standish (1982) as developed by Aoki et al. (1983), using Kinoshita's development of Andoyer's post-Newcomb precession. The numerical constants from Seidelmann (1992) are used canonically.

Conversion from B1950.0 FK4 to J2000.0 FK5 only is provided for. Conversions for different epochs and equinoxes would require additional treatment for precession, proper motion and E-terms.

In the FK4 catalog the proper motions of stars within 10 degrees of the poles do not embody differential E-terms effects and should, strictly speaking, be handled in a different manner from stars outside these regions. However, given the general lack of homogeneity of the star data available for routine astrometry, the difficulties of handling positions that may have been determined from astrometric fields spanning the polar and non- polar regions, the likelihood that the differential E-terms effect was not taken into account when allowing for proper motion in past astrometry, and the undesirability of a discontinuity in the algorithm, the decision has been made in this ERFA algorithm to include the effects of differential E-terms on the proper motions for all stars, whether polar or not. At epoch J2000.0, and measuring "on the sky" rather than in terms of RA change, the errors resulting from this simplification are less than 1 milliarcsecond in position and 1 milliarcsecond per century in proper motion.

References

Aoki, S. et al., 1983, "Conversion matrix of epoch B1950.0 FK4-based positions of stars to epoch J2000.0 positions in accordance with the new IAU resolutions". Astron.Astrophys. 128, 263-267.

Seidelmann, P.K. (ed), 1992, "Explanatory Supplement to the Astronomical Almanac", ISBN 0-935702-68-7.

Smith, C.A. et al., 1989, "The transformation of astrometric catalog systems to the equinox J2000.0". Astron.J. 97, 265.

Standish, E.M., 1982, "Conversion of positions and proper motions from B1950.0 to the IAU system at J2000.0". Astron.Astrophys., 115, 1, 20-22.

Yallop, B.D. et al., 1989, "Transformation of mean star places from FK4 B1950.0 to FK5 J2000.0 using matrices in 6-space". Astron.J. 97,

Astrometry.SOFA.starpm — Function

starpm(ras::AbstractFloat, dec::AbstractFloat, pmras::AbstractFloat, pmdec::AbstractFloat,
       plx::AbstractFloat, rvel::AbstractFloat, epoch1a::AbstractFloat, epoch1b::AbstractFloat,
       epoch2a::AbstractFloat, epoch2b::AbstractFloat)

Star proper motion: update star catalog data for space motion.

Input

ra1 – right ascension (radians), before

dec1 – declination (radians), before

pmr1 – RA proper motion (radians/year), before

pmd1 – Dec proper motion (radians/year), before

px1 – parallax (arcseconds), before

rv1 – radial velocity (km/s, +ve = receding), before

ep1a – "before" epoch, part A (Note 1)

ep1b – "before" epoch, part B (Note 1)

ep2a – "after" epoch, part A (Note 1)

ep2b – "after" epoch, part B (Note 1)

Output

ra2 – right ascension (radians), after

dec2 – declination (radians), after

pmr2 – RA proper motion (radians/year), after

pmd2 – Dec proper motion (radians/year), after

px2 – parallax (arcseconds), after

rv2 – radial velocity (km/s, +ve = receding), after

-1 = system error (should not occur) 0 = no warnings or errors 1 = distance overridden (Note 6) 2 = excessive velocity (Note 7) 4 = solution didn't converge (Note 8) else = binary logical OR of the above warnings

Note

The starting and ending TDB dates ep1a+ep1b and ep2a+ep2b are Julian Dates, apportioned in any convenient way between the two parts (A and B). For example, JD(TDB)=2450123.7 could be expressed in any of these ways, among others:

    epna          epnb

2450123.7           0.0       (JD method)
2451545.0       -1421.3       (J2000 method)
2400000.5       50123.2       (MJD method)
2450123.5           0.2       (date & time method)

The JD method is the most natural and convenient to use in cases where the loss of several decimal digits of resolution is acceptable. The J2000 method is best matched to the way the argument is handled internally and will deliver the optimum resolution. The MJD method and the date & time methods are both good compromises between resolution and convenience.

In accordance with normal star-catalog conventions, the object's right ascension and declination are freed from the effects of secular aberration. The frame, which is aligned to the catalog equator and equinox, is Lorentzian and centered on the SSB.

The proper motions are the rate of change of the right ascension and declination at the catalog epoch and are in radians per TDB Julian year.

The parallax and radial velocity are in the same frame.

Care is needed with units. The star coordinates are in radians and the proper motions in radians per Julian year, but the parallax is in arcseconds.

The RA proper motion is in terms of coordinate angle, not true angle. If the catalog uses arcseconds for both RA and Dec proper motions, the RA proper motion will need to be divided by cos(Dec) before use.

Straight-line motion at constant speed, in the inertial frame, is assumed.

An extremely small (or zero or negative) parallax is interpreted to mean that the object is on the "celestial sphere", the radius of which is an arbitrary (large) value (see the eraStarpv function for the value used). When the distance is overridden in this way, the status, initially zero, has 1 added to it.

If the space velocity is a significant fraction of c (see the constant VMAX in the function eraStarpv), it is arbitrarily set to zero. When this action occurs, 2 is added to the status.

The relativistic adjustment carried out in the eraStarpv function involves an iterative calculation. If the process fails to converge within a set number of iterations, 4 is added to the status.