API

API

EarthOrientation.EOParams — Method.

EOParams(iau1980file::String, iau2000file::String)

Parse IERS data files into a EOParams object. iau1980file and iau2000file are the paths to a 'finals.all' and a 'finals2000A.all' CSV file, respectively.

source

EarthOrientation.getdx — Method.

getdx(date; outside_range=:warn)

Get the celestial pole x-coordinate correction for a certain date in milliarcseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.getdx_err — Method.

getdx_err(date; outside_range=:warn)

Get the error in celestial pole x-coordinate correction for a certain date in milliarcseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.getdy — Method.

getdy(date; outside_range=:warn)

Get the celestial pole y-coordinate correction for a certain date in milliarcseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.getdy_err — Method.

getdy_err(date; outside_range=:warn)

Get the error in celestial pole y-coordinate correction for a certain date in milliarcseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.getdψ — Method.

getdψ(date; outside_range=:warn)

Get the ecliptic nutation correction for a certain date in milliarcseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.getdψ_err — Method.

getdψ_err(date; outside_range=:warn)

Get the error in the ecliptic nutation correction for a certain date in milliarcseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.getdϵ — Method.

getdϵ(date; outside_range=:warn)

Get the ecliptic obliquity correction for a certain date in milliarcseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.getdϵ_err — Method.

getdϵ_err(date; outside_range=:warn)

Get the error in the ecliptic obliquity correction for a certain date in milliarcseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.getlod — Method.

getlod(date; outside_range=:warn)

Get the excess length of day for a certain date in milliseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.getlod_err — Method.

getlod_err(date; outside_range=:warn)

Get the error in the excess length of day for a certain date in milliseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.getxp — Method.

getxp(date; outside_range=:warn)

Get the x-coordinate of Earth's north pole w.r.t. the CIO for a certain date in arcseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.getxp_err — Method.

getxp_err(date; outside_range=:warn)

Get the error for the x-coordinate of Earth's north pole w.r.t. the CIO for a certain date in arcseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.getyp — Method.

getyp(date; outside_range=:warn)

Get the y-coordinate of Earth's north pole w.r.t. the CIO for a certain date in arcseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.getyp_err — Method.

getyp_err(date; outside_range=:warn)

Get the error for the y-coordinate of Earth's north pole w.r.t. the CIO for a certain date in arcseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.getΔUT1 — Method.

getΔUT1(date; outside_range=:warn)

Get the difference between UTC and UT1 for a certain date in seconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.getΔUT1_err — Method.

getΔUT1_err(date; outside_range=:warn)

Get the error in the difference between UTC and UT1 for a certain date in seconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.polarmotion — Method.

polarmotion(date; outside_range=:warn)

Get the coordinates of Earth's north pole w.r.t. the CIO for a certain date in arcseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.precession_nutation00 — Method.

precession_nutation00(date; outside_range=:warn)

Get the celestial pole coordinate corrections for a certain date in milliarcseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source

EarthOrientation.precession_nutation80 — Method.

precession_nutation80(date; outside_range=:warn)

Get the ecliptic corrections for a certain date in milliarcseconds.

date can either be a DateTime object or a Julian date represented by a Float64. The outside_range argument determines what to do if no data is available for date:

:warn: The last valid value is returned and a warning will be displayed.

:nothing: The last valid value is returned.

:error: An OutOfRangeError is thrown.

source