API Reference

Reproject.interpolator — Method

interpolator(array_in, order::Int)

Returns an interpolator with the given array and order of interpolation.

Reproject.pad_edges — Method

pad_edges(array_in::Matrix{T}) where {T}

Pads a given array and creates a border with edge elements.

Reproject.parse_input_data — Method

parse_input_data(input_data::ImageHDU)
parse_input_data(input_data::Tuple{AbstractArray, WCSTransform})
parse_input_data(input_data::String, hdu_in)
parse_input_data(input_data::FITS, hdu_in)

Parse input data and returns an Array and WCS object.

Arguments

input_data: image to reproject which can be name of a FITS file, an ImageHDU or a FITS file.

hdu_in: used to set HDU to use when more than one HDU is present.

Reproject.parse_output_projection — Method

parse_output_projection(output_projection::WCSTransform, shape_out)
parse_output_projection(output_projection::ImageHDU; shape_out)
parse_output_projection(output_projection::String, hdu_number)
parse_output_projection(output_projection::FITS, hdu_number)

Parse output projection and returns a WCS object and shape of output.

Arguments

output_projection: WCS information about the image to be reprojected which can be name of a FITS file, an ImageHDU or WCSTransform.

shape_out: shape of the output image.

hdu_number: specifies HDU number when file name is given as input.

Reproject.reproject — Method

reproject(input_data, output_projection; shape_out = nothing, order = 1, hdu_in = 1, hdu_out = 1)

Reprojects image data to a new projection using interpolation.

Arguments

input_data: Image data which is being reprojected. It can be an ImageHDU, FITS object, name of a FITS file or a tuple of image matrix and WCSTransform.

output_projection: Frame in which data is reprojected. Frame can be taken from WCSTransform object, ImageHDU, FITS or name of FITS file.

shape_out: Shape of image after reprojection.

order: Order of interpolation. 0: Nearest-neighbor 1: Linear 2: Quadratic

hdu_in: Used to specify HDU number when giving input as FITS or name of FITS file.

hdu_out: Used to specify HDU number when giving output projection as FITS or name of FITS file.

Reproject.wcs_to_celestial_frame — Method

wcs_to_celestial_frame(wcs::WCSTransform)

Returns the reference frame of a WCSTransform. The reference frame supported in Julia are FK5, ICRS and Galactic.