Star Catalogs

SOFA.fk425 — Method

fk425(ra::AbstractFloat, dec::AbstractFloat, δra::AbstractFloat, δdec::AbstractFloat, plx::AbstractFloat,
      rv::AbstractFloat)

Convert B1950.0 FK4 star catalog data to J2000.0 FK5.

This function converts a star's catalog data from the old FK4 (Bessel-Newcomb) system to the later IAU 1976 FK5 (Fricke) system.

Input (all B1950.0, FK4)

ra – B1950.0 RA (rad)

dec – B1950.0 Dec (rad)

δra – B1950.0 RA proper motion (rad/trop-year)

δdec – B1950.0 Dec proper motions (rad/trop-year)

plx – parallax (arcsec)

rv – radial velocity (km/s, +ve = moving away)

Output (all J2000.0, FK5)

ra – J2000.0 RA (rad)

dec – J2000.0 Dec (rad)

δra – J2000.0 RA proper motion (rad/Jul-year)

δdec – J2000.0 Dec proper motions (rad/Jul-year)

plx – parallax (arcsec)

rv – radial velocity (km/s, +ve = moving away)

Note

The proper motions in RA are dRA/dt rather than cos(Dec)*dRA/dt, and are per year rather than per century.

The conversion is somewhat complicated, for several reasons:

. Change of standard epoch from B1950.0 to J2000.0.

. An intermediate transition date of 1984 January 1.0 TT.

. A change of precession model.

. Change of time unit for proper motion (tropical to Julian).

. FK4 positions include the E-terms of aberration, to simplify the hand computation of annual aberration. FK5 positions assume a rigorous aberration computation based on the Earth's barycentric velocity.

. The E-terms also affect proper motions, and in particular cause objects at large distances to exhibit fictitious proper motions.

The algorithm is based on Smith et al. (1989) and Yallop et al. (1989), which presented a matrix method due to Standish (1982) as developed by Aoki et al. (1983), using Kinoshita's development of Andoyer's post-Newcomb precession. The numerical constants from Seidelmann (1992) are used canonically.

Conversion from B1950.0 FK4 to J2000.0 FK5 only is provided for. Conversions for different epochs and equinoxes would require additional treatment for precession, proper motion and E-terms.

In the FK4 catalog the proper motions of stars within 10 degrees of the poles do not embody differential E-terms effects and should, strictly speaking, be handled in a different manner from stars outside these regions. However, given the general lack of homogeneity of the star data available for routine astrometry, the difficulties of handling positions that may have been determined from astrometric fields spanning the polar and non- polar regions, the likelihood that the differential E-terms effect was not taken into account when allowing for proper motion in past astrometry, and the undesirability of a discontinuity in the algorithm, the decision has been made in this SOFA algorithm to include the effects of differential E-terms on the proper motions for all stars, whether polar or not. At epoch J2000.0, and measuring "on the sky" rather than in terms of RA change, the errors resulting from this simplification are less than 1 milliarcsecond in position and 1 milliarcsecond per century in proper motion.

References

Aoki, S. et al., 1983, "Conversion matrix of epoch B1950.0 FK4-based positions of stars to epoch J2000.0 positions in accordance with the new IAU resolutions". Astron.Astrophys. 128, 263-267.

Seidelmann, P.K. (ed), 1992, "Explanatory Supplement to the Astronomical Almanac", ISBN 0-935702-68-7.

Smith, C.A. et al., 1989, "The transformation of astrometric catalog systems to the equinox J2000.0". Astron.J. 97, 265.

Standish, E.M., 1982, "Conversion of positions and proper motions from B1950.0 to the IAU system at J2000.0". Astron.Astrophys., 115, 1, 20-22.

Yallop, B.D. et al., 1989, "Transformation of mean star places from FK4 B1950.0 to FK5 J2000.0 using matrices in 6-space". Astron.J. 97,

SOFA.fk45z — Method

fk45z(ra::AbstractFloat, dec::AbstractFloat, epoch::AbstractFloat)

Convert a B1950.0 FK4 star position to J2000.0 FK5, assuming zero proper motion in the FK5 system.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

This function converts a star's catalog data from the old FK4 (Bessel-Newcomb) system to the later IAU 1976 FK5 (Fricke) system, in such a way that the FK5 proper motion is zero. Because such a star has, in general, a non-zero proper motion in the FK4 system, the function requires the epoch at which the position in the FK4 system was determined.

Input

ra – B1950.0 FK4 RA at epoch (rad)

dec – B1950.0 FK4 Dec at epoch (rad)

epoch – Besselian epoch (e.g. 1979.3)

Output

ra – J2000.0 FK5 RA (rad)

dec – J2000.0 FK5 Dec (rad)

Examples

julia> fk45z(0.01602284975382961, -0.11643479290999060, 1954.677617625256806)
(ra = 0.02719295911606862, dec = -0.1115766001565927)

Note

The epoch is strictly speaking Besselian, but if a Julian epoch is supplied the result will be affected only to a negligible extent.

The method is from Appendix 2 of Aoki et al. (1983), but using the constants of Seidelmann (1992). See the function fk425 for a general introduction to the FK4 to FK5 conversion.

Conversion from equinox B1950.0 FK4 to equinox J2000.0 FK5 only is provided for. Conversions for different starting and/or ending epochs would require additional treatment for precession, proper motion and E-terms.

In the FK4 catalog the proper motions of stars within 10 degrees of the poles do not embody differential E-terms effects and should, strictly speaking, be handled in a different manner from stars outside these regions. However, given the general lack of homogeneity of the star data available for routine astrometry, the difficulties of handling positions that may have been determined from astrometric fields spanning the polar and non- polar regions, the likelihood that the differential E-terms effect was not taken into account when allowing for proper motion in past astrometry, and the undesirability of a discontinuity in the algorithm, the decision has been made in this SOFA algorithm to include the effects of differential E-terms on the proper motions for all stars, whether polar or not. At epoch J2000.0, and measuring "on the sky" rather than in terms of RA change, the errors resulting from this simplification are less than 1 milliarcsecond in position and 1 milliarcsecond per century in proper motion.

References

Aoki, S. et al., 1983, "Conversion matrix of epoch B1950.0 FK4-based positions of stars to epoch J2000.0 positions in accordance with the new IAU resolutions". Astron.Astrophys. 128, 263-267.

Seidelmann, P.K. (ed), 1992, "Explanatory Supplement to the Astronomical Almanac", ISBN 0-935702-68-7.

SOFA.fk524 — Method

fk524(ra::AbstractFloat, dec::AbstractFloat, δra::AbstractFloat, δdec::AbstractFloat, plx::AbstractFloat,
      rv::AbstractFloat)

Convert J2000.0 FK5 star catalog data to B1950.0 FK4.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Input (all J2000.0, FK5)

ra – J2000.0 RA (rad)

dec – J2000.0 Dec (rad)

δra – J2000.0 RA proper motions (rad/Jul.yr)

δdec – J2000.0 Dec proper motions (rad/Jul.yr)

plx – parallax (arcsec)

rv – radial velocity (km/s, +ve = moving away)

Output (all B1950.0, FK4)

ra – B1950.0 RA (rad)

dec – B1950.0 Dec (rad)

δra – B1950.0 RA proper motions (rad/trop.yr)

δdec – B1950.0 Dec proper motions (rad/trop.yr)

plx – parallax (arcsec)

rv – radial velocity (km/s, +ve = moving away)

Note

The proper motions in RA are dRA/dt rather than cos(Dec)*dRA/dt, and are per year rather than per century.

The conversion is somewhat complicated, for several reasons:

. Change of standard epoch from J2000.0 to B1950.0.

. An intermediate transition date of 1984 January 1.0 TT.

. A change of precession model.

. Change of time unit for proper motion (Julian to tropical).

. FK4 positions include the E-terms of aberration, to simplify the hand computation of annual aberration. FK5 positions assume a rigorous aberration computation based on the Earth's barycentric velocity.

. The E-terms also affect proper motions, and in particular cause objects at large distances to exhibit fictitious proper motions.

The algorithm is based on Smith et al. (1989) and Yallop et al. (1989), which presented a matrix method due to Standish (1982) as developed by Aoki et al. (1983), using Kinoshita's development of Andoyer's post-Newcomb precession. The numerical constants from Seidelmann (1992) are used canonically.

In the FK4 catalog the proper motions of stars within 10 degrees of the poles do not embody differential E-terms effects and should, strictly speaking, be handled in a different manner from stars outside these regions. However, given the general lack of homogeneity of the star data available for routine astrometry, the difficulties of handling positions that may have been determined from astrometric fields spanning the polar and non- polar regions, the likelihood that the differential E-terms effect was not taken into account when allowing for proper motion in past astrometry, and the undesirability of a discontinuity in the algorithm, the decision has been made in this SOFA algorithm to include the effects of differential E-terms on the proper motions for all stars, whether polar or not. At epoch J2000.0, and measuring "on the sky" rather than in terms of RA change, the errors resulting from this simplification are less than 1 milliarcsecond in position and 1 milliarcsecond per century in proper motion.

References

Aoki, S. et al., 1983, "Conversion matrix of epoch B1950.0 FK4-based positions of stars to epoch J2000.0 positions in accordance with the new IAU resolutions". Astron.Astrophys. 128, 263-267.

Seidelmann, P.K. (ed), 1992, "Explanatory Supplement to the Astronomical Almanac", ISBN 0-935702-68-7.

Smith, C.A. et al., 1989, "The transformation of astrometric catalog systems to the equinox J2000.0". Astron.J. 97, 265.

Standish, E.M., 1982, "Conversion of positions and proper motions from B1950.0 to the IAU system at J2000.0". Astron.Astrophys., 115, 1, 20-22.

Yallop, B.D. et al., 1989, "Transformation of mean star places from FK4 B1950.0 to FK5 J2000.0 using matrices in 6-space". Astron.J. 97,

SOFA.fk52h — Method

fk52h(ra::AbstractFloat, dec::AbstractFloat, δra::AbstractFloat, δdec::AbstractFloat, plx::AbstractFloat,
      rv::AbstractFloat)

Transform FK5 (J2000.0) star data into the Hipparcos system.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Input

ra – J2000.0 FK5 RA (radians)

dec – J2000.0 FK5 Dec (radians)

δra – J2000.0 FK5 proper motion in RA (dRA/dt, rad/Jyear)

δdec – J2000.0 FK5 proper motion in Dec (dDec/dt, rad/Jyear)

plx – parallax (arcsec)

rv – radial velocity (km/s, positive = receding)

Output (all Hipparcos, epoch J2000.0)

ra – RA (radians)

dec – Dec (radians)

pmras – proper motion in RA (dRA/dt, rad/Jyear)

pmdec – proper motion in Dec (dDec/dt, rad/Jyear)

plx – parallax (arcsec)

rvel – radial velocity (km/s, positive = receding)

Note

This function transforms FK5 star positions and proper motions into the system of the Hipparcos catalog.

The proper motions in RA are dRA/dt rather than cos(Dec)*dRA/dt, and are per year rather than per century.

The FK5 to Hipparcos transformation is modeled as a pure rotation and spin; zonal errors in the FK5 catalog are not taken into account.

See also h2fk5, fk5hz, hfk5z.

References

F.Mignard & M.Froeschle, Astron.Astrophys., 354, 732-739 (2000).

SOFA.fk54z — Method

fk54z(ra::AbstractFloat, dec::AbstractFloat, epoch::AbstractFloat)

Convert a J2000.0 FK5 star position to B1950.0 FK4, assuming zero proper motion in FK5 and parallax.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Input

ra – J2000.0 FK5 RA (rad)

dec – J2000.0 FK5 Dec (rad)

epoch – Besselian epoch (e.g. 1950.0)

Output

ra – B1950.0 FK4 RA (rad) at the given epoch

dec – B1950.0 FK4 Dec (rad) at the given epoch

δra – B1950.0 FK4 proper motions in RA (rad/trop.yr)

δdec – B1950.0 FK4 proper motions in Dec (rad/trop.yr)

Note

In contrast to the fk524 function, here the FK5 proper motions, the parallax and the radial velocity are presumed zero.

This function converts a star position from the IAU 1976 FK5 (Fricke) system to the former FK4 (Bessel-Newcomb) system, for cases such as distant radio sources where it is presumed there is zero parallax and no proper motion. Because of the E-terms of aberration, such objects have (in general) non-zero proper motion in FK4, and the present function returns those fictitious proper motions.

Conversion from J2000.0 FK5 to B1950.0 FK4 only is provided for. Conversions involving other equinoxes would require additional treatment for precession.

The position returned by this function is in the B1950.0 FK4 reference system but at the given Besselian epoch. For comparison with catalogs the epoch argument will frequently be 1950.0. (In this context the distinction between Besselian and Julian epoch is insignificant.)

The RA component of the returned (fictitious) proper motion is dRA/dt rather than cos(Dec)*dRA/dt.

SOFA.fk5hip — Method

fk5hip()

FK5 to Hipparcos rotation and spin.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Output

pos – FK5 rotation matrix wrt Hipparcos (Note 2)

rot – FK5 spin vector wrt Hipparcos (Note 3)

Note

This function models the FK5 to Hipparcos transformation as a pure rotation and spin; zonal errors in the FK5 catalog are not taken into account.

The r-matrix r5h operates in the sense:

  P_Hipparcos = r5h x P_FK5

where PFK5 is a p-vector in the FK5 frame, and PHipparcos is the equivalent Hipparcos p-vector.

The r-vector s5h represents the time derivative of the FK5 to Hipparcos rotation. The units are radians per year (Julian, TDB).

References

F.Mignard & M.Froeschle, Astron.Astrophys., 354, 732-739 (2000).

SOFA.fk5hz — Method

fk5hz(ra::AbstractFloat, dec::AbstractFloat, day1::AbstractFloat, day2::AbstractFloat)

Transform an FK5 (J2000.0) star position into the system of the Hipparcos catalog, assuming zero Hipparcos proper motion.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Input

ra – FK5 RA (radians), equinox J2000.0, at date

dec – FK5 Dec (radians), equinox J2000.0, at date

day1 – TDB date (Notes 1,2)

day2 – TDB date (Notes 1,2)

Output

ra – Hipparcos RA (radians)

dec – Hipparcos Dec (radians)

Examples

julia> fk5hz(1.76779433, -0.2917517103, 2400000.5, 54479.0)
(ra = 1.7677941914644242, dec = -0.2917516001679885)

Note

This function converts a star position from the FK5 system to the Hipparcos system, in such a way that the Hipparcos proper motion is zero. Because such a star has, in general, a non-zero proper motion in the FK5 system, the function requires the date at which the position in the FK5 system was determined.

The TT date date1+date2 is a Julian Date, apportioned in any convenient way between the two arguments. For example, JD(TT)=2450123.7 could be expressed in any of these ways, among others:

   date1          date2

2450123.7           0.0       (JD method)
2451545.0       -1421.3       (J2000 method)
2400000.5       50123.2       (MJD method)
2450123.5           0.2       (date & time method)

The JD method is the most natural and convenient to use in cases where the loss of several decimal digits of resolution is acceptable. The J2000 method is best matched to the way the argument is handled internally and will deliver the optimum resolution. The MJD method and the date & time methods are both good compromises between resolution and convenience.

The FK5 to Hipparcos transformation is modeled as a pure rotation and spin; zonal errors in the FK5 catalog are not taken into account.

The position returned by this function is in the Hipparcos reference system but at date date1+date2.

See also fk52h, h2fk5, hfk5z.

References

F.Mignard & M.Froeschle, 2000, Astron.Astrophys. 354, 732-739.

SOFA.h2fk5 — Method

h2fk5(ra::AbstractFloat, dec::AbstractFloat, δra::AbstractFloat, δdec::AbstractFloat, plx::AbstractFloat,
      rv::AbstractFloat)

Transform Hipparcos star data into the FK5 (J2000.0) system.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Input

ra – J2000.0 RA (Hipparcos; radians)

dec – J2000.0 Dec (Hipparcos; radians)

δra – J2000.0 proper motion in RA (dRA/dt, rad/Jyear)

δdec – J2000.0 proper motion in Dec (dDec/dt, rad/Jyear)

plx – parallax (arcsec)

rv – radial velocity (km/s, positive = receding)

Output

ra – J2000.0 FK5 RA (radians)

dec – J2000.0 FK5 Dec (radians)

δra – J2000.0 FK5 proper motion in RA (dRA/dt, rad/Jyear)

δdec – J2000.0 FK5 proper motion in Dec (dDec/dt, rad/Jyear)

plx – parallax (arcsec)

rv – radial velocity (km/s, positive = receding)

Note

This function transforms Hipparcos star positions and proper motions into FK5 J2000.0.

The proper motions in RA are dRA/dt rather than cos(Dec)*dRA/dt, and are per year rather than per century.

The FK5 to Hipparcos transformation is modeled as a pure rotation and spin; zonal errors in the FK5 catalog are not taken into account.

See also fk52h, fk5hz, hfk5z.

References

F.Mignard & M.Froeschle, Astron.Astrophys., 354, 732-739 (2000).

SOFA.hfk5z — Method

hfk5z(ra::AbstractFloat, dec::AbstractFloat, day1::AbstractFloat, day2::AbstractFloat)

Transform a Hipparcos star position into FK5 J2000.0, assuming zero Hipparcos proper motion.

This function is part of the International Astronomical Union's SOFA (Standards of Fundamental Astronomy) software collection.

Status: support function.

Input

ra – Hipparcos RA (radians)

dec – Hipparcos Dec (radians)

day1 – TDB date (Note 1)

day2 – TDB date (Note 1)

Returned (all FK5, equinox J2000.0, date date1+date2):

ra – RA (radians)

dec – Dec (radians)

δra – RA proper motion (rad/year, Note 4)

δdec – Dec proper motion (rad/year, Note 4)

Note

The TT date date1+date2 is a Julian Date, apportioned in any convenient way between the two arguments. For example, JD(TT)=2450123.7 could be expressed in any of these ways, among others:

   date1          date2

2450123.7           0.0       (JD method)
2451545.0       -1421.3       (J2000 method)
2400000.5       50123.2       (MJD method)
2450123.5           0.2       (date & time method)

The JD method is the most natural and convenient to use in cases where the loss of several decimal digits of resolution is acceptable. The J2000 method is best matched to the way the argument is handled internally and will deliver the optimum resolution. The MJD method and the date & time methods are both good compromises between resolution and convenience.

The proper motion in RA is dRA/dt rather than cos(Dec)*dRA/dt.

The FK5 to Hipparcos transformation is modeled as a pure rotation and spin; zonal errors in the FK5 catalog are not taken into account.

It was the intention that Hipparcos should be a close approximation to an inertial frame, so that distant objects have zero proper motion; such objects have (in general) non-zero proper motion in FK5, and this function returns those fictitious proper motions.

The position returned by this function is in the FK5 J2000.0 reference system but at date date1+date2.

See also fk52h, h2fk5, fk5hz.

References

F.Mignard & M.Froeschle, 2000, Astron.Astrophys. 354, 732-739.

SOFA.starpm — Method

starpm(ra::AbstractFloat, dec::AbstractFloat, pmras::AbstractFloat, pmdec::AbstractFloat,
       plx::AbstractFloat, rvel::AbstractFloat, epoch1a::AbstractFloat, epoch1b::AbstractFloat,
       epoch2a::AbstractFloat, epoch2b::AbstractFloat)

Star proper motion: update star catalog data for space motion.

Input

ra1 – right ascension (radians), before

dec1 – declination (radians), before

pmr1 – RA proper motion (radians/year), before

pmd1 – Dec proper motion (radians/year), before

px1 – parallax (arcseconds), before

rv1 – radial velocity (km/s, +ve = receding), before

ep1a – "before" epoch, part A (Note 1)

ep1b – "before" epoch, part B (Note 1)

ep2a – "after" epoch, part A (Note 1)

ep2b – "after" epoch, part B (Note 1)

Output

ra2 – right ascension (radians), after

dec2 – declination (radians), after

pmr2 – RA proper motion (radians/year), after

pmd2 – Dec proper motion (radians/year), after

px2 – parallax (arcseconds), after

rv2 – radial velocity (km/s, +ve = receding), after

Note

The starting and ending TDB dates ep1a+ep1b and ep2a+ep2b are Julian Dates, apportioned in any convenient way between the two parts (A and B). For example, JD(TDB)=2450123.7 could be expressed in any of these ways, among others:

    epna          epnb

2450123.7           0.0       (JD method)
2451545.0       -1421.3       (J2000 method)
2400000.5       50123.2       (MJD method)
2450123.5           0.2       (date & time method)

The JD method is the most natural and convenient to use in cases where the loss of several decimal digits of resolution is acceptable. The J2000 method is best matched to the way the argument is handled internally and will deliver the optimum resolution. The MJD method and the date & time methods are both good compromises between resolution and convenience.

In accordance with normal star-catalog conventions, the object's right ascension and declination are freed from the effects of secular aberration. The frame, which is aligned to the catalog equator and equinox, is Lorentzian and centered on the SSB.

The proper motions are the rate of change of the right ascension and declination at the catalog epoch and are in radians per TDB Julian year.

The parallax and radial velocity are in the same frame.

Care is needed with units. The star coordinates are in radians and the proper motions in radians per Julian year, but the parallax is in arcseconds.

The RA proper motion is in terms of coordinate angle, not true angle. If the catalog uses arcseconds for both RA and Dec proper motions, the RA proper motion will need to be divided by cos(Dec) before use.

Straight-line motion at constant speed, in the inertial frame, is assumed.

An extremely small (or zero or negative) parallax is interpreted to mean that the object is on the "celestial sphere", the radius of which is an arbitrary (large) value (see the starpv function for the value used).

If the space velocity is a significant fraction of c (see the constant VMAX in the function starpv), it is arbitrarily set to zero.

The relativistic adjustment carried out in the starpv function involves an iterative calculation.