Datasets

SpectralFitting.jl supports a wide variety of datasets, and makes it easy to wrap your own.

For spectral fitting specifics, the main dataset type is

SpectralFitting.SpectralData — Type

SpectralData{T} <: AbstractDataset

A general spectral data structure, minimally with a Spectrum and ResponseMatrix. Optionally also includes the AncillaryResponse and a background Spectrum.

Methods

The following methods are made available through the SpectralData:

regroup!

restrict_domain!

mask_energies!

drop_channels!

drop_bad_channels!

drop_negative_channels!

normalize!

objective_units

spectrum_energy

bin_widths

subtract_background!

set_domain!

error_statistic

Constructors

The available constructors are:

SpectralData(paths::SpectralDataPaths; kwargs...)

Using SpectralDataPaths.

If the spectrum and repsonse matrix have already been loaded seperately, use

SpectralData(
    spectrum::Spectrum,
    response::ResponseMatrix;
    # try to match the domains of the response matrix to the data
    match_domains = true,
    background = nothing,
    ancillary = nothing,
)

Dataset abstraction

Datasets must define a small API to make fitting possible. The picture to have in mind when considering the different domains is as follows: the model is trying to predict the objective. It does so by taking in input domain and maps it to some output domain.

That means make_output_domain and make_objective_domain correspond to the $(X,Y)$ values of the data that the model is trying to fit, whilst the model is evaluated on the make_model_domain, which need not be the same as the output domain.

In other cases, the objective_transformer acts to transform the output of the model onto the output domain.

Mathematically, expressing the output domain $X$, the model domain $D$, the model output $M(D)$ and objective $S$, along with the transformer as $T$, then the relationship between the different domains is

\[\hat{S} = T \times M(D),\]

Both $\hat{S}$ and $S$ are defined over $X$. The various fitting operations try to find model paramters that make $\hat{S}$ and $S$ as close as possible.

SpectralFitting.AbstractDataset — Type

abstract type AbstractDataset

Abstract type for use in fitting routines. High level representation of some underlying data structures.

Fitting data is considered to have an objective and a domain. As the domain may be, for example, energy bins (high and low), or fourier frequencies (single value), the purpose of this abstraction is to provide some facility for translating between these representations for the models to fit with. This is done by checking that the AbstractDataLayout of the model and data are compatible, or at least have compatible translations.

Must implement a minimal set of accessor methods. These are paired with objective and domain parlance. Note that these functions are prefixed with make_* and not get_* to represent that there may be allocations or work going into the translation. Usage of these functions should be sparse in the interest of performance.

The arrays returned by the make_* functions must correspond to the AbstractDataLayout specified by the caller.

make_objective_variance

make_objective

make_domain_variance

make_model_domain

make_ouput_domain

Additionally there is an objective transformer that transforms the output of the model onto the output domain:

objective_transformer

Finally, to make all of the fitting for different statistical regimes work efficiently, datasets should inform which units are preferred to fit. They may also give the error statistics they prefer, and a label name primarily used to disambiguate:

preferred_units

error_statistic

make_label

SpectralFitting.make_objective_variance — Function

make_objective_variance(layout::AbstractDataLayout, dataset::AbstractDataset)

Make the variance vector associated with each objective point.

SpectralFitting.make_objective — Function

make_objective(layout::AbstractDataLayout, dataset::AbstractDataset)

Returns the array used as the target for model fitting. The array must correspond to the data AbstractDataLayout specified by the layout parameter.

In as far as it can be guarunteed, the memory in the returned array will not be mutated by any fitting procedures.

Domain for this objective should be returned by make_model_domain.

SpectralFitting.make_domain_variance — Function

make_domain_variance(layout::AbstractDataLayout, dataset::AbstractDataset)

Make the variance vector associated with the domain.

SpectralFitting.make_model_domain — Function

make_model_domain(layout::AbstractDataLayout, dataset::AbstractDataset)

Returns the array used as the domain for the modelling. This is paired with make_domain_variance

SpectralFitting.make_output_domain — Function

make_output_domain(layout::AbstractDataLayout, dataset::AbstractDataset)

Returns the array used as the output domain. That is, in cases where the model input and output map to different domains, the input domain is said to be the model domain, the input domain is said to be the model domain.

The distinction is mainly used for the purposes of simulating data and for visualising data.

Underlying data layouts

SpectralFitting.AbstractDataLayout — Type

abstract type AbstractDataLayout end

The data layout primarily concerns the relationship between the objective and the domain. It is used to work out whether a model and a dataset are fittable, and if not, whether a translation in the output of the model to the domain of the model is possible.

The following methods may be used to interrogate support:

preferred_support for inferring the preferred support of a model when multiple supports are possible.

common_support to obtain the common support of two structures

The following method is also used to define the support of a model or dataset:

supports

For cases where unit information needs to be propagated, an AbstractDataLayout can also be used to ensure the units are compatible. To query the units of a layout, use

support_units

SpectralFitting.OneToOne — Type

struct OneToOne <: AbstractDataLayout end

Indicates there is a one-to-one (injective) correspondence between each input value and each output value. That is to say

length(objective) == length(domain)

SpectralFitting.ContiguouslyBinned — Type

struct ContiguouslyBinned <: AbstractDataLayout end

Contiguously binned data layout means that the domain describes high and low bins, with the objective being the value in that bin. This means

length(objective) + 1== length(domain)

Note that the contiguous qualifer is to mean there is no gaps in the bins, and that

\[\Delta E_i = E_{i+1} - E_{i}\]

SpectralFitting.common_support — Function

common_support(x, y)

Find the common AbstractDataLayout of x and y, following the ordering of preferred_support.

SpectralFitting.preferred_support — Function

preferred_support(x)

Get the preferred AbstractDataLayout of x. If multiple supports are available, the DEFAULT_SUPPORT_ORDERING is followed:

DEFAULT_SUPPORT_ORDERING = (ContiguouslyBinned{Nothing}(nothing), OneToOne{Nothing}(nothing))

SpectralFitting.supports — Function

supports(x::Type)

Used to define whether a given type has support for a specific AbstractDataLayout. Should return a tuple of the supported layouts. This method should be implemented to express new support, not the query method.

To query, there is

supports(layout::AbstractDataLayout, x)::Bool

Example

supports(::Type{typeof(x)}) = (OneToOne(),)
@assert supports(ContiguouslyBinned(), x) == false