Astrometry.Astrometry — Module

Astrometry.jl is a set of IAU standard algorithms for calculating the time and position of celestial objects.

More information can be found on the official website of the Standards of Fundamental Astronomy.

Astrometry.iau_1976_precession — Method

iau_1976_prec(jd1, jd2)

IAU 1976 precession angles

Arguments

date1::Float64: Starting date of precession (in TDB)

date1::Float64: Ending date of precession (in TDB)

Returns

angles::Tuple{AbstractFloat}: precession angles ζ, z, θ (in radians)

The rotation matrix is Rz(-z)Ry(θ)Rz(-ζ).

The accumulated precession angles are valid for a limited time span. The absolute accuracy of the present formulation is <0.1 arcsec for 1960AD - 2040AD, <1 arcsec for 1640AD - 2360AD, and <3 arcsec for 500BC - 3000AD. The errors are >10 arcsec outside of 1200BC - 3900AD.

Astrometry.iau_1980_nutation — Method

iau_1980_nutation(date)

IAU 1980 nutation angles

Argument

date::AbstractFloat: date of nutation (in TT)

Returns

angles::Tuple{AbstractFloat}: nutation angles ψ, ϵ (in radians)

The nutation angles are with respect to the ecliptic of date.

Astrometry.iau_1980_obliquity — Method

iau_1980_obliquity(date)

IAU 1980 obliquity model angle

Arguments

date::AbstractFloat: date of obliquity (in TT)

Returns

angle::AbstractFloat: angle of obliquity at specified date (in radian)

The angle is the difference between the ecliptic and mean equator of date.

Astrometry.iau_2000_earth_rotation — Method

iau_2000_earth_rotation(date)

Earth rotation angle (IAU 2000)

Astrometry.iau_2000_position — Method

iau_2000_earth_position(date; frame=:barycenter)

Astrometry.iau_2000_precession — Method

Frame bias and precession, IAU 2000

Astrometry.iau_2006_cio_locator — Method

iau_2006_cio_locator(date, coords)

Locate the Celestial Intermediate Origin locator, s, on the equator of the Celestial Intermediate Pole, given the CIP X and Y coordinates. (IAU 2000A and 2006).

Astrometry.iau_2006_crs_trs — Method

iau_2006a_crs_trs(date)

Generate the celestial reference system (CRS) to terrestrial reference system (TRS) transform matrix given given the date, UT1, and polar motion using the IAU 2006/2000A precession nutation model.

Astrometry.iau_2006_gmst — Method

iau_2006_gmst(ut1, tt)

Greenwich Mean Sidereal Time (GMST) for the specified UT1 and TT dates.

Astrometry.iau_2006a_crs_cis — Method

iau_2006a_crs_cis(date)

Generate the celestial reference system (CRS) to celestial intermediate system (CIS) transform matrix given the CIP X & Y coordinates and the CIO locator S for the specified date using the IAU 2006/2000A precession nutation model.

Astrometry.iau_2006a_gst — Method

gst_2006a(uct, tt)

Greenwich apparent sidereal time (IAU 2000 and 2006)

Astrometry.precession_nutation — Method

precession_nutation(date; model=:IAU2006A)

Astrometry.proper_motion — Method

proper_motion(object, pmotion, parallax, rvelocity)

Astrometry.proper_motion — Method

proper_motion(object, pmotion, parallax, rvelocity, pmt, observer)

Correct coordinates for proper motion, parallax, radial velocity, and Rømer corrections.

Arguments

object::Vector{AbstractFloat}: RA and Dec of object (in radians)

propermo::Vector{AbstractFloat}: RA and Dec proper motion of object (in radians/year)

parallax::AbstractFloat: parallax of object (in arcseconds)

rvelocity::AbsractFloat: radial velocity of object (in km/sec; positive is receding)

propertim::AbstactFloat: proper motion time interval (in Julian years; at barycenter)

observer::Vector{AbstractFloat}: position of observer (in AU; from barycenter)

Returns

object::Vector{AbstractFloat}: corrected RA and Dec of object