Internals

The following types and methods are internal, and should not be considered safe for public use.

Types

Cosmology.AbstractCosmology — Type

abstract type AbstractCosmology

Abstract supertype for all cosmological models.

Cosmology.FlatLCDM — Type

struct FlatLCDM{T<:Real} <: Cosmology.AbstractFlatCosmology

ΛCDM model of the universe with $Ω_k = 0$.

Cosmology.ClosedLCDM — Type

struct ClosedLCDM{T<:Real} <: Cosmology.AbstractClosedCosmology

ΛCDM model of the universe with $Ω_k < 0$.

Cosmology.OpenLCDM — Type

struct OpenLCDM{T<:Real} <: Cosmology.AbstractOpenCosmology

ΛCDM model of the universe with $Ω_k > 0$.

Methods

Cosmology.E — Function

E(c::AbstractCosmology, z)

Dimensionless Hubble function $E(z)$ at redshift z. It's defined as

\[E(z) ≡ \frac{H(z)}{H_0} = \frac{H(z)}{(100\mathrm{km/s/Mpc}) h}\]

where $H_0 = H(z=0)$ is the Hubble parameter at the present epoch (Schneider 2015, p. 183).

Cosmology.Z — Function

Z(c::AbstractCosmology, z, nothing; kws...)
Z(c::AbstractCosmology, z₁, z₂; kws...)

The line-of-sight comoving distance contributions for comoving radial distance.

It performs the integral

\[Z = \int_{z_1}^{z_2} \frac{dz}{E(z)} = \int_{a_2}^{a_1} \frac{da}{a^2 E(a)}\]

where we can perform a change of variables with $a = 1/(1+z)$, and $dz = -da/a^2$.

If nothing is used for the second bound of integration, it defaults to z₁ = 0 (i.e., a₁ = 1).

See also

comoving_radial_dist

Cosmology.a2E — Function

a2E(c::Union{FlatLCDM,ClosedLCDM,OpenLCDM}, a)

Calculates the intermediate quantity $a^2 E(a)$. This is an internal function used to simplify computation.

Mathematical definition (for ΛCDM models):

\[a^2 E(a) = \sqrt{Ω_r + Ω_m a + Ω_k a^2 + Ω_Λ a^4}\]

where $Ω_k = 0$ for a flat cosmological model.

Cosmology.a2E — Method

a2E(c::Union{FlatWCDM,ClosedWCDM,OpenWCDM}, a)

The implementation of $a^2 E(a)$ for WCDM models.

Mathematical definition (for WCDM models):

\[a^2 E(a) = \sqrt{Ω_r + Ω_m a + Ω_k a^2 + Ω_Λ a_{de}}\]

where $a_{de} = \exp[(1 - 3 w_0 - 3 w_a) \log(a) + 3 w_a (a - 1)]$.

Cosmology.hubble_dist0 — Function

hubble_dist0(c::AbstractCosmology)

Hubble distance at redshift 0.

See also

hubble_dist

Cosmology.hubble_time0 — Function

hubble_time0(c::AbstractCosmology)

Hubble time at redshift 0.

See also

hubble_time