API/Reference

PSFModels.PSFModels

AstroLib.POLELATLONG

AstroLib.observatories

AstroLib.planets

AstroTime.TimeScales.TAI

AstroTime.TimeScales.TCB

AstroTime.TimeScales.TCG

AstroTime.TimeScales.TDB

AstroTime.TimeScales.TT

AstroTime.TimeScales.UT1

AstroImages.AstroImage

AstroImages.Comment

AstroImages.History

AstroImages.Percent

AstroImages.WCSGrid

AstroImages.Zscale

AstroLib.Observatory

AstroLib.Planet

AstroTime.AstroDates.Date

AstroTime.AstroDates.DateTime

AstroTime.AstroDates.Time

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.TAIEpoch

AstroTime.Epochs.TAIEpoch

AstroTime.Epochs.TAIEpoch

AstroTime.Epochs.TCBEpoch

AstroTime.Epochs.TCBEpoch

AstroTime.Epochs.TCBEpoch

AstroTime.Epochs.TCGEpoch

AstroTime.Epochs.TCGEpoch

AstroTime.Epochs.TCGEpoch

AstroTime.Epochs.TDBEpoch

AstroTime.Epochs.TDBEpoch

AstroTime.Epochs.TDBEpoch

AstroTime.Epochs.TTEpoch

AstroTime.Epochs.TTEpoch

AstroTime.Epochs.TTEpoch

AstroTime.Epochs.UT1Epoch

AstroTime.Epochs.UT1Epoch

AstroTime.Epochs.UT1Epoch

AstroTime.Periods.AstroPeriod

AstroTime.Periods.TimeUnit

AstroTime.TimeScales.BarycentricCoordinateTime

AstroTime.TimeScales.BarycentricDynamicalTime

AstroTime.TimeScales.GeocentricCoordinateTime

AstroTime.TimeScales.InternationalAtomicTime

AstroTime.TimeScales.TerrestrialTime

AstroTime.TimeScales.TimeScale

AstroTime.TimeScales.UniversalTime

BoxLeastSquares.BLSPeriodogram

CCDReduction.AbstractCCDData

CCDReduction.CCDData

CCDReduction.CCDData

CCDReduction.CCDData

DimensionalData.Dimensions.Dim

DimensionalData.Dimensions.LookupArrays.At

DimensionalData.Dimensions.LookupArrays.Near

DimensionalData.Dimensions.X

DimensionalData.Dimensions.Y

DimensionalData.Dimensions.Z

EarthOrientation.EOParams

FITSIO.FITS

FITSIO.FITSHeader

Photometry.Aperture.AbstractAperture

Photometry.Aperture.CircularAnnulus

Photometry.Aperture.CircularAperture

Photometry.Aperture.EllipticalAnnulus

Photometry.Aperture.EllipticalAperture

Photometry.Aperture.RectangularAnnulus

Photometry.Aperture.RectangularAperture

Photometry.Aperture.Subpixel

Photometry.Background.BackgroundInterpolator

Photometry.Background.BiweightLocationBackground

Photometry.Background.BiweightScaleRMS

Photometry.Background.IDWInterpolator

Photometry.Background.LocationEstimator

Photometry.Background.MADStdRMS

Photometry.Background.MMMBackground

Photometry.Background.RMSEstimator

Photometry.Background.SourceExtractorBackground

Photometry.Background.StdRMS

Photometry.Background.ZoomInterpolator

Photometry.Detection.PeakMesh

Photometry.Detection.SourceFinder

SkyCoords.AbstractSkyCoords

SkyCoords.FK5Coords

SkyCoords.GalCoords

SkyCoords.ICRSCoords

Transits.AbstractLimbDark

Transits.AbstractLimbDark

Transits.IntegratedLimbDark

Transits.Kipping13

Transits.Orbits.KeplerianOrbit

Transits.Orbits.SimpleOrbit

Transits.PolynomialLimbDark

Transits.QuadLimbDark

Transits.SecondaryLimbDark

WCS.WCSTransform

AstroImages.asinhstretch

AstroImages.composecolors

AstroImages.copyheader

AstroImages.header

AstroImages.implot

AstroImages.imview

AstroImages.logstretch

AstroImages.powerdiststretch

AstroImages.powstretch

AstroImages.shareheader

AstroImages.sinhstretch

AstroImages.sqrtstretch

AstroImages.squarestretch

AstroImages.wcs

AstroLib.adstring

AstroLib.airtovac

AstroLib.aitoff

AstroLib.altaz2hadec

AstroLib.baryvel

AstroLib.bprecess

AstroLib.calz_unred

AstroLib.co_aberration

AstroLib.co_nutate

AstroLib.co_refract

AstroLib.co_refract_forward

AstroLib.ct2lst

AstroLib.daycnv

AstroLib.deredd

AstroLib.eci2geo

AstroLib.eq2hor

AstroLib.eqpole

AstroLib.euler

AstroLib.flux2mag

AstroLib.gal_uvw

AstroLib.gcirc

AstroLib.geo2eci

AstroLib.geo2geodetic

AstroLib.geo2mag

AstroLib.geodetic2geo

AstroLib.get_date

AstroLib.get_juldate

AstroLib.hadec2altaz

AstroLib.helio

AstroLib.helio_jd

AstroLib.helio_rv

AstroLib.hor2eq

AstroLib.imf

AstroLib.ismeuv

AstroLib.jdcnv

AstroLib.jprecess

AstroLib.juldate

AstroLib.kepler_solver

AstroLib.lsf_rotate

AstroLib.mag2flux

AstroLib.mag2geo

AstroLib.mean_obliquity

AstroLib.month_cnv

AstroLib.moonpos

AstroLib.mphase

AstroLib.nutate

AstroLib.ordinal

AstroLib.paczynski

AstroLib.planck_freq

AstroLib.planck_wave

AstroLib.planet_coords

AstroLib.polrec

AstroLib.posang

AstroLib.precess

AstroLib.precess_cd

AstroLib.precess_xyz

AstroLib.premat

AstroLib.rad2sec

AstroLib.radec

AstroLib.recpol

AstroLib.rhotheta

AstroLib.sec2rad

AstroLib.sixty

AstroLib.sphdist

AstroLib.ten

AstroLib.tic_one

AstroLib.ticpos

AstroLib.tics

AstroLib.true_obliquity

AstroLib.trueanom

AstroLib.uvbybeta

AstroLib.vactoair

AstroLib.xyz

AstroLib.ydn2md

AstroLib.ymd2dn

AstroLib.zenpos

AstroTime.AstroDates.fractionofday

AstroTime.AstroDates.fractionofsecond

AstroTime.AstroDates.j2000

AstroTime.AstroDates.julian

AstroTime.AstroDates.julian_twopart

AstroTime.AstroDates.subsecond

AstroTime.Epochs.from_utc

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.julian_period

AstroTime.Epochs.modified_julian

AstroTime.Epochs.timescale

AstroTime.Epochs.to_utc

AstroTime.Periods.unit

AstroTime.Periods.value

AstroTime.update

Base.:-

Base.close

Base.deleteat!

Base.eltype

Base.haskey

Base.keys

Base.length

Base.length

Base.length

Base.ndims

Base.read

Base.read

Base.read!

Base.size

Base.values

Base.write

Base.write

Base.write

Base.write

BoxLeastSquares.BLS

BoxLeastSquares.autoperiod

BoxLeastSquares.model

BoxLeastSquares.model

BoxLeastSquares.params

BoxLeastSquares.periods

BoxLeastSquares.power

CCDReduction.arrays

CCDReduction.arrays

CCDReduction.ccds

CCDReduction.ccds

CCDReduction.combine

CCDReduction.crop

CCDReduction.cropview

CCDReduction.filenames

CCDReduction.filenames

CCDReduction.find_ccd

CCDReduction.fitscollection

CCDReduction.flat_correct

CCDReduction.flat_correct!

CCDReduction.getdata

CCDReduction.subtract_bias

CCDReduction.subtract_bias!

CCDReduction.subtract_dark

CCDReduction.subtract_dark!

CCDReduction.subtract_overscan

CCDReduction.subtract_overscan!

CCDReduction.trim

CCDReduction.trimview

CCDReduction.writefits

CFITSIO.bitpix_from_type

CFITSIO.cfitsio_typecode

CFITSIO.fits_clobber_file

CFITSIO.fits_close_file

CFITSIO.fits_copy_image_section

CFITSIO.fits_create_ascii_tbl

CFITSIO.fits_create_binary_tbl

CFITSIO.fits_create_diskfile

CFITSIO.fits_create_file

CFITSIO.fits_create_img

CFITSIO.fits_delete_file

CFITSIO.fits_delete_hdu

CFITSIO.fits_delete_key

CFITSIO.fits_delete_record

CFITSIO.fits_delete_rows

CFITSIO.fits_file_mode

CFITSIO.fits_file_name

CFITSIO.fits_get_coltype

CFITSIO.fits_get_hdrspace

CFITSIO.fits_get_img_size

CFITSIO.fits_get_num_hdus

CFITSIO.fits_hdr2str

CFITSIO.fits_insert_img

CFITSIO.fits_insert_rows

CFITSIO.fits_movabs_hdu

CFITSIO.fits_movnam_hdu

CFITSIO.fits_movrel_hdu

CFITSIO.fits_open_data

CFITSIO.fits_open_diskfile

CFITSIO.fits_open_file

CFITSIO.fits_open_image

CFITSIO.fits_open_table

CFITSIO.fits_read_col

CFITSIO.fits_read_keyn

CFITSIO.fits_read_keyword

CFITSIO.fits_read_pix

CFITSIO.fits_read_pixnull

CFITSIO.fits_read_record

CFITSIO.fits_read_subset

CFITSIO.fits_resize_img

CFITSIO.fits_write_col

CFITSIO.fits_write_key

CFITSIO.fits_write_null_img

CFITSIO.fits_write_pix

CFITSIO.fits_write_pixnull

CFITSIO.fits_write_record

CFITSIO.fits_write_subset

CFITSIO.libcfitsio_version

CFITSIO.type_from_bitpix

Cosmology.age

Cosmology.angular_diameter_dist

Cosmology.comoving_radial_dist

Cosmology.comoving_volume

Cosmology.comoving_volume_element

Cosmology.cosmology

Cosmology.distmod

Cosmology.lookback_time

Cosmology.luminosity_dist

Dates.day

Dates.dayofyear

Dates.hour

Dates.microsecond

Dates.millisecond

Dates.minute

Dates.month

Dates.nanosecond

Dates.now

Dates.second

Dates.second

Dates.year

Dates.yearmonthday

DimensionalData.Dimensions.dims

DimensionalData.Dimensions.refdims

EarthOrientation.getdate

EarthOrientation.getdx

EarthOrientation.getdx_err

EarthOrientation.getdy

EarthOrientation.getdy_err

EarthOrientation.getdψ

EarthOrientation.getdψ_err

EarthOrientation.getdϵ

EarthOrientation.getdϵ_err

EarthOrientation.getlod

EarthOrientation.getlod_err

EarthOrientation.getxp

EarthOrientation.getxp_err

EarthOrientation.getyp

EarthOrientation.getyp_err

EarthOrientation.getΔUT1

EarthOrientation.getΔUT1_TAI

EarthOrientation.getΔUT1_err

EarthOrientation.polarmotion

EarthOrientation.precession_nutation00

EarthOrientation.precession_nutation80

EarthOrientation.update

FITSIO.colnames

FITSIO.copy_section

FITSIO.default_header

FITSIO.fitsread

FITSIO.fitswrite

FITSIO.get_comment

FITSIO.read_header

FITSIO.read_key

FITSIO.set_comment!

FITSIO.write_key

FileIO.load

FileIO.save

LACosmic.lacosmic

PSFModels.airydisk

PSFModels.fit

PSFModels.gaussian

PSFModels.moffat

PSFModels.normal

Photometry.Aperture.photometry

Photometry.Background.estimate_background

Photometry.Background.sigma_clip

Photometry.Background.sigma_clip!

Photometry.Detection.extract_sources

SAOImageDS9.accesspoint

SAOImageDS9.bitpix_of

SAOImageDS9.bitpix_to_type

SAOImageDS9.byte_order

SAOImageDS9.connect

SAOImageDS9.draw

SAOImageDS9.get

SAOImageDS9.set

SkyCoords.offset

SkyCoords.position_angle

SkyCoords.separation

Transits.Orbits.relative_position

Transits.compute

Transits.compute

WCS.from_header

WCS.obsfix

WCS.pix_to_world

WCS.pix_to_world

WCS.pix_to_world!

WCS.to_header

WCS.world_to_pix

WCS.world_to_pix

WCS.world_to_pix!

AstroTime.@timescale

Gaussian

PSFModels.gaussian — Function

gaussian([T=Float64], point; x, y, fwhm, amp=1, theta=0, bkg=0)
gaussian([T=Float64], px, py; x, y, fwhm, amp=1, theta=0, bkg=0)

An unnormalized bivariate Gaussian distribution. The position can be specified in (x, y) coordinates as a Tuple, AbstractVector, or as separate arguments. If theta is given, the PSF will be rotated by theta degrees counter-clockwise from the x-axis. If bkg is given it will be added as a scalar to the PSF.

The fwhm can be a scalar (isotropic) or a vector/tuple (diagonal). Keep in mind that theta has no effect for isotropic distributions and is degenerate with the fwhm parameters (i.e., theta=90 is the same as reversing the fwhm tuple)

Functional form

f(x | x̂, FWHM) = exp[-4ln(2) * ||x - x̂|| / FWHM^2]

where x̂ and x are position vectors (indices) ||⋅|| represents the square-distance, and FWHM is the full width at half-maximum. If FWHM is a scalar, the Gaussian distribution will be isotropic. If FWHM is a vector or tuple, the weighting is applied along each axis (diagonal).

PSFModels.normal — Function

normal

An alias for gaussian

gauss = gaussian(x=0, y=0, fwhm=10)
psfplot(gauss, -50:50, -50:50; title="gaussian(fwhm=10)",
        colorbar_scale=:log10, clims=(1e-5, 1))

Airy Disk

PSFModels.airydisk — Function

airydisk([T=Float64], point; x, y, fwhm, ratio=0, amp=1, theta=0, bkg=0)
airydisk([T=Float64], px, py; x, y, fwhm, ratio=0, amp=1, theta=0, bkg=0)

An unnormalized Airy disk. The position can be specified in (x, y) coordinates as a Tuple, AbstractVector, or as separate arguments. If theta is given, the PSF will be rotated by theta degrees counter-clockwise from the x-axis. If bkg is given it will be added as a scalar to the PSF.

The fwhm can be a scalar (isotropic) or a vector/tuple (diagonal). Keep in mind that theta has no effect for isotropic distributions and is degenerate with the fwhm parameters (i.e., theta=90 is the same as reversing the fwhm tuple)

If ratio is supplied, this will be the Airy pattern for a centrally-obscured aperture (e.g., a Newtonian telescope). This has a slightly expanded functional form, and in general the central Airy disk will be smaller and the first Airy ring will be brighter.

Functional form

The Airy disk is a distribution over the radius r (the square-Euclidean distance)

f(x | x̂, FWHM) = [ 2J₁(q) / q ]^2

where J₁ is the first-order Bessel function of the first kind and

q ≈ π * r * D/ λ ≈ π * r / (0.973 * FWHM)

If user a non-zero central obscuration via ratio, the functional form becomes

f(x | x̂, FWHM, ϵ) = [ 2J₁(q) / q - 2ϵJ₁(ϵq) / q ]^2 / (1 - ϵ^2)^2

where ϵ is the ratio (0 ≤ ϵ < 1).

airy = airydisk(x=0, y=0, fwhm=10)
psfplot(airy, -50:50, -50:50; title="airydisk(fwhm=10)",
        colorbar_scale=:log10, clims=(1e-5, 1))

airy_obscured = airydisk(x=0, y=0, fwhm=10, ratio=0.3)
psfplot(airy_obscured, -50:50, -50:50; title="airydisk(fwhm=10, ratio=0.3)",
        colorbar_scale=:log10, clims=(1e-5, 1))

Moffat

PSFModels.moffat — Function

moffat([T=Float64], point; x, y, fwhm, alpha=1, amp=1, theta=0, bkg=0)
moffat([T=Float64], px, py; x, y, fwhm, alpha=1, amp=1, theta=0, bkg=0)

Two dimensional Moffat model. The position can be specified in (x, y) coordinates as a Tuple, AbstractVector, or as separate arguments. If theta is given, the PSF will be rotated by theta degrees counter-clockwise from the x-axis. If bkg is given it will be added as a scalar to the PSF.

The fwhm can be a scalar (isotropic) or a vector/tuple (diagonal). Keep in mind that theta has no effect for isotropic distributions and is degenerate with the fwhm parameters (i.e., theta=90 is the same as reversing the fwhm tuple)

Functional form

f(x | x̂, FWHM, α) = A / (1 + ||x - x̂|| / (FWHM / 2)^2)^α

where x̂ and x are position vectors (indices) ||⋅|| represents the square-distance, and FWHM is the full width at half-maximum. If FWHM is a vector or tuple, the weighting is applied along each axis.

moff = moffat(x=0, y=0, fwhm=10)
psfplot(moff, -50:50, -50:50; title="moffat(fwhm=10)",
        colorbar_scale=:log10, clims=(1e-5, 1))

moff2 = moffat(x=0, y=0, fwhm=10, alpha=2)
psfplot(moff2, -50:50, -50:50; title="moffat(fwhm=10, alpha=2)",
        colorbar_scale=:log10, clims=(1e-5, 1))

Comparison

xs = range(0, 50, length=1000)
plot(
    xs, [gauss.(xs, 0) airy.(xs, 0) moff.(xs, 0)],
    label=["gaussian" "airydisk" "moffat"], yscale=:log10,
    xlabel="x", ylabel="I", ylims=(1e-5, 1)
)

Fitters

PSFModels.fit — Function

PSFModels.fit(model, params, image, inds=axes(image); func_kwargs=(;), loss=abs2, maxfwhm=Inf, alg=LBFGS(), kwargs...)

Fit a PSF model (model) defined by the given params as a named tuple of the parameters to fit and their default values. This model is fit to the data in image at the specified inds (by default, the entire array). To pass extra keyword arguments to the model (i.e., to "freeze" a parameter), pass them in a named tuple to func_kwargs. The default loss function is the chi-squared loss, which uses the the square of the difference (i.e., the L2 norm). You can change this to the L1 norm, for example, by passing loss=abs. The maximum FWHM can be set with maxfwhm as a number or tuple.

Additional keyword arguments, as well as the fitting algorithm alg, are passed to Optim.optimize. By default we use forward-mode auto-differentiation (AD) to derive Jacobians for the LBFGS optimization algorithm. Refer to the Optim.jl documentation for more information.

Choosing parameters

The fit function is very powerful because it gives you a great amount of flexibility in the way you fit your models. To demonstrate this, let's start with a simple isotropic gaussian.

model = gaussian
# match params to arguments of PSF
params = (x=20, y=20, fwhm=3, amp=1)

Note that params can follow any order

params = (amp=1, x=20, y=20, fwhm=3)

Now, to extend this interface to the bivariate PSF case, where fwhm is two values, all you need to do is use a tuple or vector

params = (x=20, y=20, fwhm=(3, 3))

and, again, the order does not matter

model = moffat
params = (alpha=1, x=20, y=20, fwhm=3, amp=10)

Fitting a PSF

After selecting your model and parameters, fitting data is easy

P = (x=12, y=13, fwhm=13.2, amp=0.1)
psf = gaussian.(CartesianIndicies(1:25, 1:15); P...)

params, synthpsf = PSFModels.fit(gaussian, P, psf)

here params is a named tuple of the best fitting parameters. It will not include any fixed parameters.

synthpsf is the best-fitting model, for direct comparison with the input data.

psf_fit = synthpsf.(CartesianIndicies(psf))

To alter parameters without fitting them (i.e., "freeze" them) use func_kwargs

P = (x=12, y=13, fwhm=(12.4, 13.2), amp=0.1)
func_kwargs = (alpha=2)
params, synthpsf = PSFModels.fit(moffat, P, psf; func_kwargs)