API/Reference

Index

AstroLib.POLELATLONG

AstroLib.observatories

AstroLib.planets

AstroTime.TimeScales.TAI

AstroTime.TimeScales.TCB

AstroTime.TimeScales.TCG

AstroTime.TimeScales.TDB

AstroTime.TimeScales.TT

AstroTime.TimeScales.UT1

AstroImages.AstroImage

AstroImages.Comment

AstroImages.History

AstroImages.Percent

AstroImages.WCSGrid

AstroImages.Zscale

AstroLib.Observatory

AstroLib.Planet

AstroTime.AstroDates.Date

AstroTime.AstroDates.DateTime

AstroTime.AstroDates.Time

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.TAIEpoch

AstroTime.Epochs.TAIEpoch

AstroTime.Epochs.TAIEpoch

AstroTime.Epochs.TCBEpoch

AstroTime.Epochs.TCBEpoch

AstroTime.Epochs.TCBEpoch

AstroTime.Epochs.TCGEpoch

AstroTime.Epochs.TCGEpoch

AstroTime.Epochs.TCGEpoch

AstroTime.Epochs.TDBEpoch

AstroTime.Epochs.TDBEpoch

AstroTime.Epochs.TDBEpoch

AstroTime.Epochs.TTEpoch

AstroTime.Epochs.TTEpoch

AstroTime.Epochs.TTEpoch

AstroTime.Epochs.UT1Epoch

AstroTime.Epochs.UT1Epoch

AstroTime.Epochs.UT1Epoch

AstroTime.Periods.AstroPeriod

AstroTime.Periods.TimeUnit

AstroTime.TimeScales.BarycentricCoordinateTime

AstroTime.TimeScales.BarycentricDynamicalTime

AstroTime.TimeScales.GeocentricCoordinateTime

AstroTime.TimeScales.InternationalAtomicTime

AstroTime.TimeScales.TerrestrialTime

AstroTime.TimeScales.TimeScale

AstroTime.TimeScales.UniversalTime

BoxLeastSquares.BLSPeriodogram

DimensionalData.Dimensions.Dim

DimensionalData.Dimensions.LookupArrays.At

DimensionalData.Dimensions.LookupArrays.Near

DimensionalData.Dimensions.X

DimensionalData.Dimensions.Y

DimensionalData.Dimensions.Z

EarthOrientation.EOParams

FITSIO.FITS

FITSIO.FITSHeader

Photometry.Aperture.AbstractAperture

Photometry.Aperture.CircularAnnulus

Photometry.Aperture.CircularAperture

Photometry.Aperture.EllipticalAnnulus

Photometry.Aperture.EllipticalAperture

Photometry.Aperture.RectangularAnnulus

Photometry.Aperture.RectangularAperture

Photometry.Aperture.Subpixel

Photometry.Background.BackgroundInterpolator

Photometry.Background.BiweightLocationBackground

Photometry.Background.BiweightScaleRMS

Photometry.Background.IDWInterpolator

Photometry.Background.LocationEstimator

Photometry.Background.MADStdRMS

Photometry.Background.MMMBackground

Photometry.Background.RMSEstimator

Photometry.Background.SourceExtractorBackground

Photometry.Background.StdRMS

Photometry.Background.ZoomInterpolator

Photometry.Detection.PeakMesh

Photometry.Detection.SourceFinder

SkyCoords.AbstractSkyCoords

SkyCoords.EclipticCoords

SkyCoords.FK5Coords

SkyCoords.GalCoords

SkyCoords.ICRSCoords

WCS.WCSTransform

AstroImages.asinhstretch

AstroImages.composecolors

AstroImages.copyheader

AstroImages.header

AstroImages.implot

AstroImages.imview

AstroImages.logstretch

AstroImages.powerdiststretch

AstroImages.powstretch

AstroImages.shareheader

AstroImages.sinhstretch

AstroImages.sqrtstretch

AstroImages.squarestretch

AstroImages.wcs

AstroLib.adstring

AstroLib.airtovac

AstroLib.aitoff

AstroLib.altaz2hadec

AstroLib.baryvel

AstroLib.bprecess

AstroLib.calz_unred

AstroLib.co_aberration

AstroLib.co_nutate

AstroLib.co_refract

AstroLib.co_refract_forward

AstroLib.ct2lst

AstroLib.daycnv

AstroLib.deredd

AstroLib.eci2geo

AstroLib.eq2hor

AstroLib.eqpole

AstroLib.euler

AstroLib.flux2mag

AstroLib.gal_uvw

AstroLib.gcirc

AstroLib.geo2eci

AstroLib.geo2geodetic

AstroLib.geo2mag

AstroLib.geodetic2geo

AstroLib.get_date

AstroLib.get_juldate

AstroLib.hadec2altaz

AstroLib.helio

AstroLib.helio_jd

AstroLib.helio_rv

AstroLib.hor2eq

AstroLib.imf

AstroLib.ismeuv

AstroLib.jdcnv

AstroLib.jprecess

AstroLib.juldate

AstroLib.kepler_solver

AstroLib.lsf_rotate

AstroLib.mag2flux

AstroLib.mag2geo

AstroLib.mean_obliquity

AstroLib.month_cnv

AstroLib.moonpos

AstroLib.mphase

AstroLib.nutate

AstroLib.ordinal

AstroLib.paczynski

AstroLib.planck_freq

AstroLib.planck_wave

AstroLib.planet_coords

AstroLib.polrec

AstroLib.posang

AstroLib.precess

AstroLib.precess_cd

AstroLib.precess_xyz

AstroLib.premat

AstroLib.rad2sec

AstroLib.radec

AstroLib.recpol

AstroLib.rhotheta

AstroLib.sec2rad

AstroLib.sixty

AstroLib.sphdist

AstroLib.sunpos

AstroLib.ten

AstroLib.tic_one

AstroLib.ticpos

AstroLib.tics

AstroLib.true_obliquity

AstroLib.trueanom

AstroLib.uvbybeta

AstroLib.vactoair

AstroLib.xyz

AstroLib.ydn2md

AstroLib.ymd2dn

AstroLib.zenpos

AstroTime.AstroDates.fractionofday

AstroTime.AstroDates.fractionofsecond

AstroTime.AstroDates.j2000

AstroTime.AstroDates.julian

AstroTime.AstroDates.julian_twopart

AstroTime.AstroDates.subsecond

AstroTime.Epochs.from_utc

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.julian_period

AstroTime.Epochs.modified_julian

AstroTime.Epochs.timescale

AstroTime.Epochs.to_utc

AstroTime.Periods.unit

AstroTime.Periods.value

AstroTime.update

Base.:-

Base.close

Base.deleteat!

Base.eltype

Base.haskey

Base.keys

Base.length

Base.length

Base.length

Base.ndims

Base.read

Base.read

Base.read!

Base.size

Base.values

Base.write

Base.write

Base.write

Base.write

BoxLeastSquares.BLS

BoxLeastSquares.autoperiod

BoxLeastSquares.model

BoxLeastSquares.model

BoxLeastSquares.params

BoxLeastSquares.periods

BoxLeastSquares.power

CFITSIO.bitpix_from_type

CFITSIO.cfitsio_typecode

CFITSIO.fits_clobber_file

CFITSIO.fits_close_file

CFITSIO.fits_copy_image_section

CFITSIO.fits_create_ascii_tbl

CFITSIO.fits_create_binary_tbl

CFITSIO.fits_create_diskfile

CFITSIO.fits_create_file

CFITSIO.fits_create_img

CFITSIO.fits_delete_file

CFITSIO.fits_delete_hdu

CFITSIO.fits_delete_key

CFITSIO.fits_delete_record

CFITSIO.fits_delete_rows

CFITSIO.fits_file_mode

CFITSIO.fits_file_name

CFITSIO.fits_get_coltype

CFITSIO.fits_get_hdrspace

CFITSIO.fits_get_img_size

CFITSIO.fits_get_num_hdus

CFITSIO.fits_hdr2str

CFITSIO.fits_insert_img

CFITSIO.fits_insert_rows

CFITSIO.fits_movabs_hdu

CFITSIO.fits_movnam_hdu

CFITSIO.fits_movrel_hdu

CFITSIO.fits_open_data

CFITSIO.fits_open_diskfile

CFITSIO.fits_open_file

CFITSIO.fits_open_image

CFITSIO.fits_open_table

CFITSIO.fits_read_col

CFITSIO.fits_read_keyn

CFITSIO.fits_read_keyword

CFITSIO.fits_read_pix

CFITSIO.fits_read_pixnull

CFITSIO.fits_read_record

CFITSIO.fits_read_subset

CFITSIO.fits_resize_img

CFITSIO.fits_write_col

CFITSIO.fits_write_key

CFITSIO.fits_write_null_img

CFITSIO.fits_write_pix

CFITSIO.fits_write_pixnull

CFITSIO.fits_write_record

CFITSIO.fits_write_subset

CFITSIO.libcfitsio_version

CFITSIO.type_from_bitpix

Cosmology.age

Cosmology.angular_diameter_dist

Cosmology.comoving_radial_dist

Cosmology.comoving_volume

Cosmology.comoving_volume_element

Cosmology.cosmology

Cosmology.distmod

Cosmology.lookback_time

Cosmology.luminosity_dist

Dates.day

Dates.dayofyear

Dates.hour

Dates.microsecond

Dates.millisecond

Dates.minute

Dates.month

Dates.nanosecond

Dates.now

Dates.second

Dates.second

Dates.year

Dates.yearmonthday

DimensionalData.Dimensions.dims

DimensionalData.Dimensions.refdims

EarthOrientation.getdate

EarthOrientation.getdx

EarthOrientation.getdx_err

EarthOrientation.getdy

EarthOrientation.getdy_err

EarthOrientation.getdψ

EarthOrientation.getdψ_err

EarthOrientation.getdϵ

EarthOrientation.getdϵ_err

EarthOrientation.getlod

EarthOrientation.getlod_err

EarthOrientation.getxp

EarthOrientation.getxp_err

EarthOrientation.getyp

EarthOrientation.getyp_err

EarthOrientation.getΔUT1

EarthOrientation.getΔUT1_TAI

EarthOrientation.getΔUT1_err

EarthOrientation.polarmotion

EarthOrientation.precession_nutation00

EarthOrientation.precession_nutation80

EarthOrientation.update

FITSIO.colnames

FITSIO.copy_section

FITSIO.default_header

FITSIO.fitsread

FITSIO.fitswrite

FITSIO.get_comment

FITSIO.read_header

FITSIO.read_key

FITSIO.set_comment!

FITSIO.write_key

FileIO.load

FileIO.save

LACosmic.lacosmic

LombScargle.M

LombScargle.autofrequency

LombScargle.bootstrap

LombScargle.fap

LombScargle.fap

LombScargle.fapinv

LombScargle.fapinv

LombScargle.findmaxfreq

LombScargle.findmaxperiod

LombScargle.findmaxpower

LombScargle.freq

LombScargle.freqpower

LombScargle.lombscargle

LombScargle.lombscargle

LombScargle.model

LombScargle.period

LombScargle.periodpower

LombScargle.plan

LombScargle.power

LombScargle.prob

LombScargle.probinv

Photometry.Aperture.photometry

Photometry.Background.estimate_background

Photometry.Background.sigma_clip

Photometry.Background.sigma_clip!

Photometry.Detection.extract_sources

SAOImageDS9.accesspoint

SAOImageDS9.bitpix_of

SAOImageDS9.bitpix_to_type

SAOImageDS9.byte_order

SAOImageDS9.connect

SAOImageDS9.draw

SAOImageDS9.get

SAOImageDS9.set

SkyCoords.offset

SkyCoords.position_angle

SkyCoords.separation

WCS.from_header

WCS.obsfix

WCS.pix_to_world

WCS.pix_to_world

WCS.pix_to_world!

WCS.to_header

WCS.world_to_pix

WCS.world_to_pix

WCS.world_to_pix!

AstroTime.@timescale

Types

SkyCoords.AbstractSkyCoords — Type

The supertype for all sky coordinate systems.

SkyCoords.ICRSCoords — Type

ICRSCoords(ra, dec)

International Celestial Reference System

This is the current standard adopted by the International Astronomical Union notably due to its high level of accuracy compared to standard equatorial coordinate systems. What sets this apart from FK5Coords is that it is completely defined using extragalactic radio sources rather than a geocentric frame, which means the reference frame will not change due to Earth's motion.

Coordinates

ra - Right ascension in radians (0, 2π)

dec - Declination in radians (-π/2, π/2)

SkyCoords.GalCoords — Type

GalCoords(l, b)

Galactic Coordinate System

This coordinate system is defined based on the projection of the Milky Way galaxy onto our celestial sphere, with (0, 0) being approximately the center of our galaxy.

Coordinates

l - Galactic longitude in radians (-π, π)

b - Galactic latitude in radians (-π/2, π/2)

SkyCoords.FK5Coords — Type

FK5Coords{equinox}(ra, dec)

Equatorial Coordinate System

This coordinate system maps the celestial sphere based on a geocentric observer. Historically the oldest, this coordinate system has been shown to be inaccurate due to its definitions based on the Earth, which has long-scale precession causing the reference frame to change. Because of this, an equinox must be provided (typically 2000, commonly known as J2000) which defines the reference frame.

Coordinates

ra - Right ascension in radians (0, 2π)

dec - Declination in radians (-π/2, π/2)

SkyCoords.EclipticCoords — Type

EclipticCoords{equinox}(lon, lat)

Ecliptic Coordinate System

This coordinate system is geocentric with the ecliptic plane as the xy-plane with x oriented according to the equinox specified by equinox.

Coordinates

lon - Longitude in radians (0, 2π)

lat - Latitude in radians (-π/2, π/2)

Conversion

To convert between types, there are three (equivalent) methods of doing so.

julia> c1 = ICRSCoords(0., 0.)
ICRSCoords{Float64}(0.0, 0.0)

using convert

julia> convert(GalCoords, c1)
GalCoords{Float64}(1.6814027872278692, -1.0504884034813007)

using constructors

julia> GalCoords(c1)
GalCoords{Float64}(1.6814027872278692, -1.0504884034813007)

using |>

julia> c1 |> GalCoords
GalCoords{Float64}(1.6814027872278692, -1.0504884034813007)

Functions

SkyCoords.separation — Function

separation(c1::AbstractSkyCoords, c2::AbstractSkyCoords) -> distance

Return angular separation between two sky coordinates, in radians.

The angular separation is calculated using the Vincenty formula, which is slightly more complex and computationally expensive than some alternatives, but is stable at at all distances, including the poles and antipodes.

SkyCoords.position_angle — Function

position_angle(c1::AbstractSkyCoords, c2::AbstractSkyCoords) -> angle

Return position angle between two sky coordinates, in positive radians.

Examples

julia> c1 = ICRSCoords(0, 0); c2 = ICRSCoords(deg2rad(1), 0);

julia> position_angle(c1, c2) |> rad2deg
90.0

SkyCoords.offset — Function

offset(::AbstractSkyCoords, separation, pa) -> coordinate

Offset a coordinate by a given angular separation, separation, in radians and position angle, pa, in radians.

Uses the sine and cosine rules in spherical coordinates with corrections for the antipodes. Returns a sky coordinate of the same type as input.

Examples

julia> c1 = ICRSCoords(0, 0);

julia> c2 = offset(c1, deg2rad(1), deg2rad(90))
ICRSCoords{Float64}(0.017453292519943295, 1.0686516840418957e-18)

julia> offset(c1, c2) .|> rad2deg
(1.0, 90.0)

See Also

separation, position_angle

offset(::AbstractSkyCoords, AbstractSkyCoords) -> angle, angle

Return the separation and position angle in radians between two sky coordinates.

Examples

julia> c1 = ICRSCoords(0, 0); c2 = ICRSCoords(deg2rad(1), 0);

julia> offset(c1, c2) .|> rad2deg
(1.0, 90.0)

See Also

separation, position_angle