API/Reference

Index

PSFModels.PSFModels

AstroLib.POLELATLONG

AstroLib.observatories

AstroLib.planets

AstroTime.TimeScales.TAI

AstroTime.TimeScales.TCB

AstroTime.TimeScales.TCG

AstroTime.TimeScales.TDB

AstroTime.TimeScales.TT

AstroTime.TimeScales.UT1

AstroImages.AstroImage

AstroImages.Comment

AstroImages.History

AstroImages.Percent

AstroImages.WCSGrid

AstroImages.Zscale

AstroLib.Observatory

AstroLib.Planet

AstroTime.AstroDates.Date

AstroTime.AstroDates.DateTime

AstroTime.AstroDates.Time

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.Epoch

AstroTime.Epochs.TAIEpoch

AstroTime.Epochs.TAIEpoch

AstroTime.Epochs.TAIEpoch

AstroTime.Epochs.TCBEpoch

AstroTime.Epochs.TCBEpoch

AstroTime.Epochs.TCBEpoch

AstroTime.Epochs.TCGEpoch

AstroTime.Epochs.TCGEpoch

AstroTime.Epochs.TCGEpoch

AstroTime.Epochs.TDBEpoch

AstroTime.Epochs.TDBEpoch

AstroTime.Epochs.TDBEpoch

AstroTime.Epochs.TTEpoch

AstroTime.Epochs.TTEpoch

AstroTime.Epochs.TTEpoch

AstroTime.Epochs.UT1Epoch

AstroTime.Epochs.UT1Epoch

AstroTime.Epochs.UT1Epoch

AstroTime.Periods.AstroPeriod

AstroTime.Periods.TimeUnit

AstroTime.TimeScales.BarycentricCoordinateTime

AstroTime.TimeScales.BarycentricDynamicalTime

AstroTime.TimeScales.GeocentricCoordinateTime

AstroTime.TimeScales.InternationalAtomicTime

AstroTime.TimeScales.TerrestrialTime

AstroTime.TimeScales.TimeScale

AstroTime.TimeScales.UniversalTime

BoxLeastSquares.BLSPeriodogram

CCDReduction.AbstractCCDData

CCDReduction.CCDData

CCDReduction.CCDData

CCDReduction.CCDData

DimensionalData.Dimensions.Dim

DimensionalData.Dimensions.LookupArrays.At

DimensionalData.Dimensions.LookupArrays.Near

DimensionalData.Dimensions.X

DimensionalData.Dimensions.Y

DimensionalData.Dimensions.Z

EarthOrientation.EOParams

FITSIO.FITS

FITSIO.FITSHeader

Photometry.Aperture.AbstractAperture

Photometry.Aperture.CircularAnnulus

Photometry.Aperture.CircularAperture

Photometry.Aperture.EllipticalAnnulus

Photometry.Aperture.EllipticalAperture

Photometry.Aperture.RectangularAnnulus

Photometry.Aperture.RectangularAperture

Photometry.Aperture.Subpixel

Photometry.Background.BackgroundInterpolator

Photometry.Background.BiweightLocationBackground

Photometry.Background.BiweightScaleRMS

Photometry.Background.IDWInterpolator

Photometry.Background.LocationEstimator

Photometry.Background.MADStdRMS

Photometry.Background.MMMBackground

Photometry.Background.RMSEstimator

Photometry.Background.SourceExtractorBackground

Photometry.Background.StdRMS

Photometry.Background.ZoomInterpolator

Photometry.Detection.PeakMesh

Photometry.Detection.SourceFinder

SkyCoords.AbstractSkyCoords

SkyCoords.FK5Coords

SkyCoords.GalCoords

SkyCoords.ICRSCoords

Transits.AbstractLimbDark

Transits.AbstractLimbDark

Transits.IntegratedLimbDark

Transits.Kipping13

Transits.Orbits.KeplerianOrbit

Transits.Orbits.SimpleOrbit

Transits.PolynomialLimbDark

Transits.QuadLimbDark

Transits.SecondaryLimbDark

WCS.WCSTransform

AstroImages.asinhstretch

AstroImages.composecolors

AstroImages.copyheader

AstroImages.header

AstroImages.implot

AstroImages.imview

AstroImages.logstretch

AstroImages.powerdiststretch

AstroImages.powstretch

AstroImages.shareheader

AstroImages.sinhstretch

AstroImages.sqrtstretch

AstroImages.squarestretch

AstroImages.wcs

AstroLib.adstring

AstroLib.airtovac

AstroLib.aitoff

AstroLib.altaz2hadec

AstroLib.baryvel

AstroLib.bprecess

AstroLib.calz_unred

AstroLib.co_aberration

AstroLib.co_nutate

AstroLib.co_refract

AstroLib.co_refract_forward

AstroLib.ct2lst

AstroLib.daycnv

AstroLib.deredd

AstroLib.eci2geo

AstroLib.eq2hor

AstroLib.eqpole

AstroLib.euler

AstroLib.flux2mag

AstroLib.gal_uvw

AstroLib.gcirc

AstroLib.geo2eci

AstroLib.geo2geodetic

AstroLib.geo2mag

AstroLib.geodetic2geo

AstroLib.get_date

AstroLib.get_juldate

AstroLib.hadec2altaz

AstroLib.helio

AstroLib.helio_jd

AstroLib.helio_rv

AstroLib.hor2eq

AstroLib.imf

AstroLib.ismeuv

AstroLib.jdcnv

AstroLib.jprecess

AstroLib.juldate

AstroLib.kepler_solver

AstroLib.lsf_rotate

AstroLib.mag2flux

AstroLib.mag2geo

AstroLib.mean_obliquity

AstroLib.month_cnv

AstroLib.moonpos

AstroLib.mphase

AstroLib.nutate

AstroLib.ordinal

AstroLib.paczynski

AstroLib.planck_freq

AstroLib.planck_wave

AstroLib.planet_coords

AstroLib.polrec

AstroLib.posang

AstroLib.precess

AstroLib.precess_cd

AstroLib.precess_xyz

AstroLib.premat

AstroLib.rad2sec

AstroLib.radec

AstroLib.recpol

AstroLib.rhotheta

AstroLib.sec2rad

AstroLib.sixty

AstroLib.sphdist

AstroLib.ten

AstroLib.tic_one

AstroLib.ticpos

AstroLib.tics

AstroLib.true_obliquity

AstroLib.trueanom

AstroLib.uvbybeta

AstroLib.vactoair

AstroLib.xyz

AstroLib.ydn2md

AstroLib.ymd2dn

AstroLib.zenpos

AstroTime.AstroDates.fractionofday

AstroTime.AstroDates.fractionofsecond

AstroTime.AstroDates.j2000

AstroTime.AstroDates.julian

AstroTime.AstroDates.julian_twopart

AstroTime.AstroDates.subsecond

AstroTime.Epochs.from_utc

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.getoffset

AstroTime.Epochs.julian_period

AstroTime.Epochs.modified_julian

AstroTime.Epochs.timescale

AstroTime.Epochs.to_utc

AstroTime.Periods.unit

AstroTime.Periods.value

AstroTime.update

Base.:-

Base.close

Base.deleteat!

Base.eltype

Base.haskey

Base.keys

Base.length

Base.length

Base.length

Base.ndims

Base.read

Base.read

Base.read!

Base.size

Base.values

Base.write

Base.write

Base.write

Base.write

BoxLeastSquares.BLS

BoxLeastSquares.autoperiod

BoxLeastSquares.model

BoxLeastSquares.model

BoxLeastSquares.params

BoxLeastSquares.periods

BoxLeastSquares.power

CCDReduction.arrays

CCDReduction.arrays

CCDReduction.ccds

CCDReduction.ccds

CCDReduction.combine

CCDReduction.crop

CCDReduction.cropview

CCDReduction.filenames

CCDReduction.filenames

CCDReduction.find_ccd

CCDReduction.fitscollection

CCDReduction.flat_correct

CCDReduction.flat_correct!

CCDReduction.getdata

CCDReduction.subtract_bias

CCDReduction.subtract_bias!

CCDReduction.subtract_dark

CCDReduction.subtract_dark!

CCDReduction.subtract_overscan

CCDReduction.subtract_overscan!

CCDReduction.trim

CCDReduction.trimview

CCDReduction.writefits

CFITSIO.bitpix_from_type

CFITSIO.cfitsio_typecode

CFITSIO.fits_clobber_file

CFITSIO.fits_close_file

CFITSIO.fits_copy_image_section

CFITSIO.fits_create_ascii_tbl

CFITSIO.fits_create_binary_tbl

CFITSIO.fits_create_diskfile

CFITSIO.fits_create_file

CFITSIO.fits_create_img

CFITSIO.fits_delete_file

CFITSIO.fits_delete_hdu

CFITSIO.fits_delete_key

CFITSIO.fits_delete_record

CFITSIO.fits_delete_rows

CFITSIO.fits_file_mode

CFITSIO.fits_file_name

CFITSIO.fits_get_coltype

CFITSIO.fits_get_hdrspace

CFITSIO.fits_get_img_size

CFITSIO.fits_get_num_hdus

CFITSIO.fits_hdr2str

CFITSIO.fits_insert_img

CFITSIO.fits_insert_rows

CFITSIO.fits_movabs_hdu

CFITSIO.fits_movnam_hdu

CFITSIO.fits_movrel_hdu

CFITSIO.fits_open_data

CFITSIO.fits_open_diskfile

CFITSIO.fits_open_file

CFITSIO.fits_open_image

CFITSIO.fits_open_table

CFITSIO.fits_read_col

CFITSIO.fits_read_keyn

CFITSIO.fits_read_keyword

CFITSIO.fits_read_pix

CFITSIO.fits_read_pixnull

CFITSIO.fits_read_record

CFITSIO.fits_read_subset

CFITSIO.fits_resize_img

CFITSIO.fits_write_col

CFITSIO.fits_write_key

CFITSIO.fits_write_null_img

CFITSIO.fits_write_pix

CFITSIO.fits_write_pixnull

CFITSIO.fits_write_record

CFITSIO.fits_write_subset

CFITSIO.libcfitsio_version

CFITSIO.type_from_bitpix

Cosmology.age

Cosmology.angular_diameter_dist

Cosmology.comoving_radial_dist

Cosmology.comoving_volume

Cosmology.comoving_volume_element

Cosmology.cosmology

Cosmology.distmod

Cosmology.lookback_time

Cosmology.luminosity_dist

Dates.day

Dates.dayofyear

Dates.hour

Dates.microsecond

Dates.millisecond

Dates.minute

Dates.month

Dates.nanosecond

Dates.now

Dates.second

Dates.second

Dates.year

Dates.yearmonthday

DimensionalData.Dimensions.dims

DimensionalData.Dimensions.refdims

EarthOrientation.getdate

EarthOrientation.getdx

EarthOrientation.getdx_err

EarthOrientation.getdy

EarthOrientation.getdy_err

EarthOrientation.getdψ

EarthOrientation.getdψ_err

EarthOrientation.getdϵ

EarthOrientation.getdϵ_err

EarthOrientation.getlod

EarthOrientation.getlod_err

EarthOrientation.getxp

EarthOrientation.getxp_err

EarthOrientation.getyp

EarthOrientation.getyp_err

EarthOrientation.getΔUT1

EarthOrientation.getΔUT1_TAI

EarthOrientation.getΔUT1_err

EarthOrientation.polarmotion

EarthOrientation.precession_nutation00

EarthOrientation.precession_nutation80

EarthOrientation.update

FITSIO.colnames

FITSIO.copy_section

FITSIO.default_header

FITSIO.fitsread

FITSIO.fitswrite

FITSIO.get_comment

FITSIO.read_header

FITSIO.read_key

FITSIO.set_comment!

FITSIO.write_key

FileIO.load

FileIO.save

LACosmic.lacosmic

PSFModels.airydisk

PSFModels.fit

PSFModels.gaussian

PSFModels.moffat

PSFModels.normal

Photometry.Aperture.photometry

Photometry.Background.estimate_background

Photometry.Background.sigma_clip

Photometry.Background.sigma_clip!

Photometry.Detection.extract_sources

SAOImageDS9.accesspoint

SAOImageDS9.bitpix_of

SAOImageDS9.bitpix_to_type

SAOImageDS9.byte_order

SAOImageDS9.connect

SAOImageDS9.draw

SAOImageDS9.get

SAOImageDS9.set

SkyCoords.offset

SkyCoords.position_angle

SkyCoords.separation

Transits.Orbits.relative_position

Transits.compute

Transits.compute

WCS.from_header

WCS.obsfix

WCS.pix_to_world

WCS.pix_to_world

WCS.pix_to_world!

WCS.to_header

WCS.world_to_pix

WCS.world_to_pix

WCS.world_to_pix!

AstroTime.@timescale

Light Curves

Transits.AbstractLimbDark — Type

AbstractLimbDark

A limb dark law need only need to implement compute(::Law, b, r) to extend the limb darkening interface.

See also

compute

Transits.AbstractLimbDark — Method

(::AbstractLimbDark)(b, r)

An alias for calling compute

Examples

julia> ld = PolynomialLimbDark([0.4, 0.26]);

julia> ld(0, 0.01)
0.9998785437247428

Transits.PolynomialLimbDark — Type

PolynomialLimbDark(u::AbstractVector)

Polynomial limb darkening using analytical integrals. The length of the u vector is equivalent to the order of polynomial used; e.g., [0.2, 0.3] corresponds to quadratic limb darkening.

Mathematical form

\[I(\mu) \propto 1 - u_1(1-\mu) - u_2(1-\mu)^2 - \dots - u_N(1-\mu)^N\]

which is equivalent to the series

\[I(\mu) \propto -\sum_{i=0}^N{u_i(1-\mu)^i}\]

with the definition $u_0 \equiv -1$.

Examples

u = [0.4, 0.26] # quadratic and below is 100% analytical
ld = PolynomialLimbDark(u)
ld(0.1, 0.01)

# output
0.9998787880717668

u2 = vcat(u, ones(12) ./ 12)
ld2 = PolynomialLimbDark(u2)
ld2(0.1, 0.01)

# output
0.9998740059086433

References

Agol, Luger, Foreman-Mackey (2020)

"Analytic Planetary Transit Light Curves and Derivatives for Stars with Polynomial Limb Darkening"

Luger et al. (2019)

"starry: Analytic Occultation Light Curves"

Transits.QuadLimbDark — Type

QuadLimbDark(u::AbstractVector)

A specialized implementation of PolynomialLimbDark with a maximum of two terms (quadratic form). This has a completely closed-form solution without any numerical integration. This means there are no intermediate allocations and reduced numerical error.

Mathematical form

\[I(\mu) \propto 1 - u_1(1-\mu) - u_2(1-\mu)^2\]

Higher-order terms

Higher-order terms will be ignored; no error will be thrown

Examples

ld = QuadLimbDark(Float64[]) # constant term only

b = [0, 1, 2] # impact parameter
r = 0.01 # radius ratio
ld.(b, r)

# output
3-element Vector{Float64}:
 0.9999
 0.9999501061035608
 1.0

ld = QuadLimbDark([0.4, 0.26]) # max two terms
ld.(b, r)

# output
3-element Vector{Float64}:
 0.9998785437247428
 0.999974726693709
 1.0

References

See references for PolynomialLimbDark

Transits.IntegratedLimbDark — Type

IntegratedLimbDark(limbdark; N=21, basis=:legendre)
IntegratedLimbDark(u; kwargs...)

Computes the time-averaged flux in the middle of an exposure by wrapping a limb darkening law limbdark with a quadrature scheme. For each time step t, N extra points are super-sampled from t-texp/2 to t+texp/2and the time-averaged flux is calculated via quadrature.

If a set of limb darkening coefficients, u, is provided, a PolynomialLimbDark law will be used by default.

Mathematical form

\[\bar{F}(t) = \frac{1}{\Delta t}\int_{t-\Delta t / 2}^{t+\Delta t / 2}{F(t')dt'}\]

where $F$ is the wrapped limb darkening law and $\Delta t$ is the exposure time.

Quadrature

The integration is approximated via Guassian quadrature

\[\frac{1}{\Delta t} \int{F(t')dt'} \approx \frac12\sum_i^N{w_i * F(\frac{\Delta t}{2}\xi_i + t)}\]

where the weights w_i and nodes ξ_i are defined by the given quadrature rule. The nodes are defined by evaluating orthogonal polynomials N times between -1 and 1. Notice the change of interval required to go from the natural bounds of the orthogonal polynomial basis, -1, 1, to the range defined by the exposure time.

The following bases are available from FastGaussQuadrature.jl. In addition, a function can be passed which calculates nodes, weights = f(N).

:legendre - Legendre polynomial base on the open (-1, 1)

:radau - Legendre polynomial base on the semi-open [-1, 1) interval

:lobatto - Legendre polynomial base on the closed [-1, 1] interval

Transits.SecondaryLimbDark — Type

SecondaryLimbDark(primary::AbstractLimbDark,
                  secondary::AbstractLimbDark; 
                  brightness_ratio=1)
SecondaryLimbDark(u_p::AbstractVector, u_s=u_p; kwargs...)

Compose two limb darkening laws together to add a secondary eclipse. If vectors of coefficients are provided, laws will automatically be constructed using PolynomialLimbDark. The surface brightness ratio is given in terms of the host; e.g., if the companion is half as bright as the host, the ratio would be 0.5.

Interface

SecondaryLimbDark only works with an orbit, since the companion's reference frame needs to be calculated. This means you can't call it using an impact parameter like ld(b, r) directly.

Mathematical form

\[f(t, r) = \frac{2f_p(t, r) + \eta r^2 f_s(t', r')}{1 + f_p(t, r) + \eta r^2 f_s(t', r')}\]

where $f_p$ is to the primary flux, $f_s$ is to the secondary flux, and $\eta$ is the surface brightness ratio. $t'$ and $r'$ correspond to the time and radius ratio from the companion's reference frame.

Examples

# equal size and limb darkening
r = 1.0
u = [0.4, 0.26]
# companion is 1/10 as bright
brightness_ratio = 0.1
ld = SecondaryLimbDark(u; brightness_ratio)
orbit = SimpleOrbit(period=2, duration=0.5)
fp = ld(orbit, 0, r) # primary egress
fs = ld(orbit, 1, r) # secondary egress

fp ≈ brightness_ratio * fs

# output
true

Transits.compute — Function

compute(::AbstractLimbDark, b, r; kwargs...)

Compute the relative flux for the given impact parameter b and radius ratio r. The impact parameter is unitless. The radius ratio is given in terms of the host; e.g., if the companion is half the size of the host, r=0.5.

Transits.compute — Method

compute(::AbstractLimbDark, orbit::AbstractOrbit, t, r)

Compute the relative flux by calculating the impact parameter at time t from the given orbit. The time needs to be compatible with the period of the orbit, nominally in days.

Examples

julia> ld = PolynomialLimbDark([0.4, 0.26]);

julia> orbit = SimpleOrbit(period=3, duration=1);

julia> ld(orbit, 0, 0.1) # primary egress
0.9878664434953113

julia> ld(orbit, 0.1, 0.1) # 0.1 d
0.9879670695533511

this works effortlessly with libraries like Unitful.jl

julia> using Unitful

julia> orbit = SimpleOrbit(period=3u"d", duration=3u"hr");

julia> ld(orbit, 0u"d", 0.1)
0.9878664434953113

Gradients

Gradients and jacobians are integrated directly into ChainRules.jl via frules and rrules. For most users, this just means using AD libraries like ForwardDiff.jl and Zygote.jl is effortless and fast.

using Transits
using Zygote

lightcurve(X) = compute(PolynomialLimbDark(X[3:end]), X[1], X[2])
grad(X) = lightcurve'(X) # Zygote gradient
grad([0.1, 0.1, 0.4, 0.26])

# output
4-element Vector{Float64}:
  0.0004972185834858653
 -0.2419262730830416
 -0.0048107583897073185
 -0.0024501564976671724

To help demonstrate the logic behind these chain rules, here we derive a simple gradient function manually.

using ChainRules

u_n = [0.4, 0.26]
μ = 0.1
ror = 0.1
X0 = [μ, ror, u_n...]

function gradr(X)
    ld, ld_pullback = rrule(PolynomialLimbDark, X[3:end])
    f, f_pullback = rrule(compute, ld, X[1], X[2])

    f̄ = one(eltype(X))
    _, l̄d, b̄, r̄ = f_pullback(f̄)
    _, ū_n = ld_pullback(l̄d)
    return [b̄, r̄, ū_n...]
end

gradr([0.1, 0.1, 0.4, 0.26])

# output
4-element Vector{Float64}:
  0.0004972185834858653
 -0.2419262730830416
 -0.0048107583897073185
 -0.0024501564976671724

For the most granular support for gradients and jacobians, peer into the depths of polynomial/poly-grad.jl and polynomial/quad-grad.jl. These functions are not part of the public API and are not guaranteed any stability according to semantic versioning.

Orbits

Transits.Orbits.SimpleOrbit — Type

SimpleOrbit(; period, duration, t0=0, b=0.0)

Circular orbit parameterized by the basic observables of a transiting system.

Parameters

period - The orbital period of the planets, nominally in days

duration - The duration of the transit, similar units as period.

t0 - The midpoint time of the reference transit, similar units as period

b - The impact parameter of the orbit, unitless

Transits.Orbits.KeplerianOrbit — Type

KeplerianOrbit(; kwargs...)

Keplerian orbit parameterized by the basic observables of a transiting 2-body system.

Parameters

period/P – The orbital period of the planet [d].

t0/t_0 – The midpoint time of the reference transit [d].

tp/t_p – The time of periastron [d].

duration/τ/T – The transit duration [d].

a – The semi-major axis [R⊙].

aR_star/aRs – The ratio of the semi-major axis to star radius.

R_planet/Rp – The radius of the planet [R⊙].

R_star/Rs – The radius of the star [R⊙].

rho_star/ρ_star – The spherical star density [M⊙/R⊙³].

r/RpRs – The ratio of the planet radius to star radius.

b – The impact parameter, bounded between 0 ≤ b ≤ 1.

ecc/e – The eccentricity of the closed orbit, bounded between 0 ≤ ecc < 1.

incl – The inclination of the orbital plane relative to the axis perpendicular to the reference plane [rad]

omega/ω – The argument of periapsis [rad].

cos_omega/cos_ω – The cosine of the argument of periapsis.

sin_omega/sin_ω – The sine of the argument of periapsis.

Omega/Ω – The longitude of the ascending node [rad].

M_planet/Mp – The mass of the planet [M⊙].

M_star/Ms – The mass of the star [M⊙].

Valid combinations

The following flowchart can be used to determine which parameters can define a KeplerianOrbit:

The period or a must be given. If both given, then neither M_star or rho_star can be defined because the stellar density is now implied.

Only incl or b can be given.

If ecc is given, then omega must also be given.

If no stellar parameters are given, the central body is assumed to be the Sun. If only rho_star is given, then R_star is defined to be 1 solar radius. Otherwise, at most two of M_star, R_star, and rho_star can be given.

Either t0 or tp must be given, but not both.

Transits.Orbits.relative_position — Function

relative_position(::AbstractOrbit, t)

The relative position, [x, y, z], of the companion compared to the host at time t. In other words, this is the vector pointing from the host to the companion along the line of sight. Nominally, the units of this distance are relative to the host's radius. For example, a distance of 2 is 2 stellar radii.

Distributions

Transits.Kipping13 — Type

Kipping13()

A non-informative prior for two-parameter limb-darkening coefficients using triangular sampling (Kipping 2013).

Examples

julia> using Random; rng = Random.seed!(10);

julia> rand(rng, Kipping13())
2-element Vector{Float64}:
 0.24716310305467298
 0.08836997249298882

julia> rand(rng, Kipping13(), 5)
2×5 Matrix{Float64}:
 0.0664907  0.124817   1.00732    0.806902  0.74165
 0.520411   0.222718  -0.389412  -0.314755  0.0768429

References

Kipping (2013)

"Efficient, uninformative sampling of limb darkening coefficients for two-parameter laws"