DustExtinction.jl

DustExtinction.jl

DustExtinction.jl

Installation

Usage

Reference/API

Installation

From the REPL, press ] to enter Pkg mode

(v 1.1) pkg> add DustExtinction
[...]

julia> using DustExtinction

Usage

julia> using DustExtinction

Color laws

julia> ccm89(4000., 3.1)
1.4645557029425842

These laws can be applied across higher dimension arrays using the . operator

julia> ccm89.([4000., 5000.], 3.1)
2-element Array{Float64,1}:
 1.4645557029425842
 1.122246878899302

If you want to apply total extinction $A_V$ it's as simple as multiplcation

julia> a_v=0.3
0.3

julia> a_v * ccm89(4000., 3.1)
0.43936671088277524

Advanced Examples

The color laws also have built-in support for uncertainties using Measurements.jl.

julia> using Measurements

julia> ccm89.([4000. ± 10.5, 5000. ± 10.2], 3.1)
2-element Array{Measurement{Float64},1}:
 1.4646 ± 0.0033
 1.1222 ± 0.003

and also support units via Unitful.jl and its subsidiaries.

julia> using Unitful, UnitfulAstro

julia> mags = ccm89.([4000u"angstrom", 0.5u"μm"], 3.1)
2-element Array{Gain{Unitful.LogInfo{:Magnitude,10,-2.5},:?,Float64},1}:
 1.4645557029425837 mag
  1.122246878899302 mag

julia> fluxes = ones(2)u"erg/s/cm^2"
2-element Array{Quantity{Float64,𝐌*𝐓^-3,Unitful.FreeUnits{(erg, cm^-2, s^-1),𝐌*𝐓^-3,nothing}},1}:
 1.0 erg cm^-2 s^-1
 1.0 erg cm^-2 s^-1

julia> reddened = fluxes .* mags
2-element Array{Quantity{Float64,𝐌*𝐓^-3,Unitful.FreeUnits{(erg, cm^-2, s^-1),𝐌*𝐓^-3,nothing}},1}:
  0.2595241150526157 erg cm^-2 s^-1
 0.35571423768348603 erg cm^-2 s^-1

julia> reddened ≈ @. fluxes * 10 ^ (-0.4 * ustrip(mags))
true

You can even combine the two above to get some really nice workflows exploiting all Julia has to offer! This example would shows how you could redden some OIR observational data with uncertainties in the flux density.

julia> using Random, Measurements, Unitful, UnitfulAstro

julia> wave = range(0.3, 1.0, length=5)u"μm"
0.3 μm:0.175 μm:1.0 μm

julia> err = randn(MersenneTwister(0), length(wave))u"erg/s/cm^2"
5-element Array{Quantity{Float64,𝐌*𝐓^-3,Unitful.FreeUnits{(erg, cm^-2, s^-1),𝐌*𝐓^-3,nothing}},1}:
   0.6791074260357777 erg cm^-2 s^-1
   0.8284134829000359 erg cm^-2 s^-1
  -0.3530074003005963 erg cm^-2 s^-1
 -0.13485387193052173 erg cm^-2 s^-1
   0.5866170746331097 erg cm^-2 s^-1

julia> flux = @.(ustrip(u"erg/s/cm^2", 300u"erg/s*cm^2" / wave^4) ± ustrip(err))u"erg/s/cm^2"
5-element Array{Quantity{Measurement{Float64},𝐌*𝐓^-3,Unitful.FreeUnits{(erg, cm^-2, s^-1),𝐌*𝐓^-3,nothing}},1}:
  3.7037037037037034e20 ± 0.68 erg cm^-2 s^-1
   5.893140783143162e19 ± 0.83 erg cm^-2 s^-1
 1.6806134238997232e19 ± -0.35 erg cm^-2 s^-1
  6.475979428646598e18 ± -0.13 erg cm^-2 s^-1
                 3.0e18 ± 0.59 erg cm^-2 s^-1

julia> λ_m = @. 0.25 * ccm89(wave, 3.1)
5-element Array{Gain{Unitful.LogInfo{:Magnitude,10,-2.5},:?,Float64},1}:
  0.4545391743613166 mag
  0.3000686820198816 mag
  0.2068677378008488 mag
 0.14069431885378056 mag
 0.10099999999999998 mag

julia> flux .* λ_m
5-element Array{Quantity{Measurement{Float64},𝐌*𝐓^-3,Unitful.FreeUnits{(erg, cm^-2, s^-1),𝐌*𝐓^-3,nothing}},1}:
 2.4368038160281967e20 ± 0.45 erg cm^-2 s^-1
  4.470121662545148e19 ± 0.63 erg cm^-2 s^-1
  1.389059128986988e19 ± 0.29 erg cm^-2 s^-1
  5.688892575540536e18 ± 0.12 erg cm^-2 s^-1
  2.733513699128758e18 ± 0.53 erg cm^-2 s^-1

Dust maps

julia> ENV["SFD98_DIR"] = "/home/user/data/dust"

# download maps (once)
julia> download_sfd98()

julia> dustmap = SFD98Map()
SFD98Map("/home/user/data/dust")

julia> ebv_galactic(dustmap, 0.1, 0.1)
0.793093095733043

julia> ebv_galactic(dustmap, [0.1, 0.2], [0.1, 0.2])
2-element Array{Float64,1}:
 0.793093
 0.539507

Reference/API

Extinction Laws

DustExtinction.ccm89 — Function.

ccm89(wave::Real, r_v::Real=3.1)
ccm89(wave::Quantity, r_v::Real=3.1)

Clayton, Cardelli and Mathis (1989) dust law.

Returns the extinction in magnitudes at the given wavelength(s) wave (in Angstroms), relative to the extinction at 5494.5 Angstroms. The parameter r_v changes the shape of the function. A typical value for the Milky Way is 3.1. An error is raised for wavelength values outside the range of support, 1000. to 33333.33 Angstroms.

References

[1] Cardelli, Clayton and Mathis (1989)

source

DustExtinction.od94 — Function.

od94(wave::Real, r_v::Real=3.1)
od94(wave::Quantity, r_v::Real=3.1)

O'Donnell (1994) dust law.

This is identical to the Clayton, Cardelli and Mathis (1989) dust law, except that different coefficients are used in the optical (3030.3 to 9090.9 Angstroms). Returns the extinction in magnitudes at the given wavelength(s) wave (in Angstroms), relative to the extinction at 5494.5 Angstroms. The parameter r_v changes the shape of the function. A typical value for the Milky Way is 3.1. An error is raised for wavelength values outside the range of support, 1000. to 33333.33 Angstroms.

References

[1] O'Donnell (1994)

source

DustExtinction.cal00 — Function.

cal00(wave::Real, r_v::Real=3.1)
cal00(wave::Quantity, r_v::Real=3.1)

Calzetti et al. (2000) Dust law.

Calculate the magnitudes for given wavelengths wave in Angstrom. Wavelength support is 0.12 to 2.2 microns (error will be thrown if out of this range). Accepts selective extinction r_v parameter with default set to Milky Way average of 3.1.

References

[1] Calzetti et al. (2000)

source

Dust Maps

DustExtinction.download_sfd98 — Function.

download_sfd98([destdir])

Download the Schlegel, Finkbeiner and Davis (1998) dust maps to the given directory. If the directory is ommitted, the SFD98_DIR environment variable is used as the destination directory.

source

DustExtinction.SFD98Map — Type.

SFD98Map([mapdir])

Schlegel, Finkbeiner and Davis (1998) dust map. mapdir should be a directory containing the two FITS files defining the map, SFD_dust_4096_[ngp,sgp].fits. If mapdir is omitted, the SFD98_DIR environment variable is used. Internally, this type keeps the FITS files defining the map open, speeding up repeated queries for E(B-V) values.

source

DustExtinction.ebv_galactic — Function.

ebv_galactic(dustmap::SFD98Map, l::Real, b::Real)
ebv_galactic(dustmap::SFD98Map, l::Vector{<:Real}, b::Vector{<:Real})

Get E(B-V) value from a SFD98Map instance at galactic coordinates (l, b), given in radians. l and b may be Vectors. Uses bilinear interpolation between pixel values.

source