API Documentation

FileIO.load — Function

load(filename) loads the contents of a formatted file, trying to infer the format from filename and/or magic bytes in the file (see query).

load(strm) loads from an IOStream or similar object. In this case, there is no filename extension, so we rely on the magic bytes for format identification.

load(File{format"PNG"}(filename)) specifies the format directly, and bypasses the format query.

load(Stream{format"PNG"}(io)) specifies the format directly, and bypasses the format query.

load(f; options...) passes keyword arguments on to the loader.

FileIO.save — Function

save(filename, data...) saves the contents of a formatted file, trying to infer the format from filename.

save(Stream{format"PNG"}(io), data...) specifies the format directly, and bypasses the format query.

save(File{format"PNG"}(filename), data...) specifies the format directly, and bypasses the format query.

save(f, data...; options...) passes keyword arguments on to the saver.

Missing docstring.

Missing docstring for AstroImage. Check Documenter's build log for details.

Missing docstring.

Missing docstring for imview. Check Documenter's build log for details.

Missing docstring.

Missing docstring for implot. Check Documenter's build log for details.

DimensionalData.Dimensions.dims — Function

dims(x, [dims::Tuple]) => Tuple{Vararg{Dimension}}
dims(x, dim) => Dimension

Return a tuple of Dimensions for an object, in the order that matches the axes or columns of the underlying data.

dims can be Dimension, Dimension types, or Symbols for Dim{Symbol}.

The default is to return nothing.

DimensionalData.Dimensions.refdims — Function

refdims(x, [dims::Tuple]) => Tuple{Vararg{Dimension}}
refdims(x, dim) => Dimension

Reference dimensions for an array that is a slice or view of another array with more dimensions.

slicedims(a, dims) returns a tuple containing the current new dimensions and the new reference dimensions. Refdims can be stored in a field or disgarded, as it is mostly to give context to plots. Ignoring refdims will simply leave some captions empty.

The default is to return an empty Tuple().

Missing docstring.

Missing docstring for Comment. Check Documenter's build log for details.

Missing docstring.

Missing docstring for History. Check Documenter's build log for details.

WCS.pix_to_world — Function

pix_to_world(wcs, pixcoords)

Convert the array of pixel coordinates pixcoords to world coordinates according to the WCSTransform wcs. pixcoords should be a 2-d array where pixcoords[:, i] is the i-th set of coordinates, or a 1-d array representing a single set of coordinates.

The return value is the same shape as pixcoords.

WCS.world_to_pix — Function

world_to_pix(wcs, worldcoords)

Convert the array of world coordinates worldcoords to pixel coordinates according to the WCSTransform wcs. worldcoords is a 2-d array where worldcoords[:, i] is the i-th set of coordinates, or a 1-d array representing a single set of coordinates.

The return value is the same size as worldcoords.

DimensionalData.Dimensions.X — Type

X <: XDim

X(val=:)

X Dimension. X <: XDim <: IndependentDim

Example:

xdim = X(2:2:10)
# Or
val = A[X(1)]
# Or
mean(A; dims=X)

DimensionalData.Dimensions.Y — Type

Y <: YDim

Y(val=:)

Y Dimension. Y <: YDim <: DependentDim

Example:

ydim = Y(['a', 'b', 'c'])
# Or
val = A[Y(1)]
# Or
mean(A; dims=Y)

DimensionalData.Dimensions.Z — Type

Z <: ZDim

Z(val=:)

Z Dimension. Z <: ZDim <: Dimension

Example:

zdim = Z(10:10:100)
# Or
val = A[Z(1)]
# Or
mean(A; dims=Z)

DimensionalData.Dimensions.Dim — Type

Dim{S}(val=:)

A generic dimension. For use when custom dims are required when loading data from a file. Can be used as keyword arguments for indexing.

Dimension types take precedence over same named Dim types when indexing with symbols, or e.g. creating Tables.jl keys.

using DimensionalData

dim = Dim{:custom}(['a', 'b', 'c'])

# output

Dim{:custom} Char['a', 'b', 'c']

DimensionalData.Dimensions.LookupArrays.At — Type

At <: IntSelector

At(x, atol, rtol)
At(x; atol=nothing, rtol=nothing)

Selector that exactly matches the value on the passed-in dimensions, or throws an error. For ranges and arrays, every intermediate value must match an existing value - not just the end points.

x can be any value or Vector of values.

atol and rtol are passed to isapprox. For Numberrtol will be set to Base.rtoldefault, otherwise nothing, and wont be used.

Example

using DimensionalData

A = DimArray([1 2 3; 4 5 6], (X(10:10:20), Y(5:7)))
A[X(At(20)), Y(At(6))]

# output

5

DimensionalData.Dimensions.LookupArrays.Near — Type

Near <: IntSelector

Near(x)

Selector that selects the nearest index to x.

With Points this is simply the index values nearest to the x, however with Intervals it is the interval center nearest to x. This will be offset from the index value for Start and End loci.

Example

using DimensionalData

A = DimArray([1 2 3; 4 5 6], (X(10:10:20), Y(5:7)))
A[X(Near(23)), Y(Near(5.1))]

# output
4

Missing docstring.

Missing docstring for header. Check Documenter's build log for details.

Missing docstring.

Missing docstring for wcs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for WCSGrid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for composecolors. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Zscale. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Percent. Check Documenter's build log for details.

Missing docstring.

Missing docstring for logstretch. Check Documenter's build log for details.

Missing docstring.

Missing docstring for powstretch. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sqrtstretch. Check Documenter's build log for details.

Missing docstring.

Missing docstring for squarestretch. Check Documenter's build log for details.

Missing docstring.

Missing docstring for asinhstretch. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sinhstretch. Check Documenter's build log for details.

Missing docstring.

Missing docstring for powerdiststretch. Check Documenter's build log for details.

Missing docstring.

Missing docstring for copyheader. Check Documenter's build log for details.

Missing docstring.

Missing docstring for shareheader. Check Documenter's build log for details.