API/Reference

PSFModels.PSFModels

PSFModels.AiryDisk

PSFModels.Gaussian

PSFModels.Moffat

PSFModels.Normal

PSFModels.PSFModel

PSFModels.PSFModel — Type

PSFModels.PSFModel{T} <: AbstractMatrix{T}

Abstract type for PSF models.

In general, all PSFModels have a set of pre-determined axes (the size is set upon creation) but they are lazy. That is, no memory is allocated and the values are calculated on the fly.

Interface

The interface to define a model is as follows (for an example model Model)

method	description

Model()	constructor(s)

Base.size(m::Model)	size, necessary for AbstractArray interface

Base.axes(m::Model)	axes, necessary for AbstractArray interface

(m::Model)(point::AbstractVector)	evaluate the model at the point in 2d space (x, y)

browsing through the implementation of PSFModels.Gaussian should give a good idea of how to create a model

Gaussian

PSFModels.Gaussian — Type

PSFModels.Gaussian([T=Float64]; fwhm, x=0, y=0, amp=1, maxsize=3, extent=maxsize .* fwhm)
PSFModels.Gaussian([T=Float64]; fwhm, pos, amp=1, maxsize=3, extent=maxsize .* fwhm)
PSFModels.Gaussian([T=Float64]; fwhm, r, theta, amp=1, maxsize=3, extent=maxsize .* fwhm)

An unnormalized bivariate Gaussian distribution. The position can be specified in (x, y) coordinates as a Tuple, AbstractVector, or as separate arguments. By default the model is placed at the origin. The position can also be given as a polar coordinate using r/ρ and theta/θ, optionally centered around origin.

The fwhm can be a scalar (isotropic) or a vector/tuple (diagonal). For efficient calculations, we recommend using StaticArrays. Here, maxsize is a multiple of the fwhm, and can be given as a scalar or as a tuple for each axis. The extent defines the bounding box for the model and is used for the default rendering size.

Functional form

f(x | x̂, FWHM) = exp[-4ln(2) * ||x - x̂|| / FWHM^2]

where x̂ and x are position vectors (indices) ||⋅|| represents the square-distance, and FWHM is the full width at half-maximum. If FWHM is a scalar, the Gaussian distribution will be isotropic. If FWHM is a vector or tuple, the weighting is applied along each axis (diagonal).

PSFModels.Normal — Type

PSFModels.Normal

An alias for PSFModels.Gaussian

model = Gaussian(fwhm=10)
plot(model; title="Gaussian(fwhm=10)")

Airy Disk

PSFModels.AiryDisk — Type

PSFModels.AiryDisk([T=Float64]; fwhm, x=0, y=0, amp=1, maxsize=3, extent=maxsize .* fwhm)
PSFModels.AiryDisk([T=Float64]; fwhm, pos, amp=1, maxsize=3, extent=maxsize .* fwhm)
PSFModels.AiryDisk([T=Float64]; fwhm, r, theta, amp=1, maxsize=3, extent=maxsize .* fwhm)

An unnormalized Airy disk. The position can be specified in (x, y) coordinates as a Tuple, AbstractVector, or as separate arguments. By default the model is placed at the origin. The position can also be given as a polar coordinate using r/ρ and theta/θ, optionally centered around origin.

The fwhm can be a scalar (isotropic) or a vector/tuple (diagonal). For efficient calculations, we recommend using StaticArrays. Here, maxsize is a multiple of the fwhm, and can be given as a scalar or as a tuple for each axis. The extent defines the bounding box for the model and is used for the default rendering size.

Functional form

The Airy disk is a distribution over the radius r (the square-Euclidean distance)

f(x | x̂, FWHM) = [ 2J₁(q) / q ]^2

where J₁ is the first order Bessel function of the first kind and

q ≈ π * r / (0.973 * FWHM)

model = AiryDisk(fwhm=10)
plot(model; title="AiryDisk(fwhm=10)")

Moffat

PSFModels.Moffat — Type

PSFModels.Moffat([T=Float64]; fwhm, x=0, y=0, amp=1, alpha=1, maxsize=3, extent=maxsize .* fwhm)
PSFModels.Moffat([T=Float64]; fwhm, pos, amp=1, alpha=1, maxsize=3, extent=maxsize .* fwhm)
PSFModels.Moffat([T=Float64]; fwhm, r, theta, amp=1, alpha=1, maxsize=3, extent=maxsize .* fwhm)

Two dimensional Moffat model. The position can be specified in (x, y) coordinates as a Tuple, AbstractVector, or as separate arguments. By default the model is placed at the origin. The position can also be given as a polar coordinate using r/ρ and theta/θ, optionally centered around origin.

The fwhm can be a scalar (isotropic) or a vector/tuple (diagonal). For efficient calculations, we recommend using StaticArrays. Here, maxsize is a multiple of the fwhm, and can be given as a scalar or as a tuple for each axis. The extent defines the bounding box for the model and is used for the default rendering size.

Functional form

f(x | x̂, FWHM, α) = A / (1 + ||x - x̂|| / (FWHM / 2)^2)^α

where x̂ and x are position vectors (indices) ||⋅|| represents the square-distance, and FWHM is the full width at half-maximum. If FWHM is a vector or tuple, the weighting is applied along each axis.

model = Moffat(fwhm=10)
plot(model; title="Moffat(fwhm=10)")