API/Reference

For now, here is a list of all documented functions.

Index

CCDReduction.combine

CCDReduction.combine

CCDReduction.crop

CCDReduction.crop

CCDReduction.cropview

CCDReduction.flat_correct

CCDReduction.flat_correct

CCDReduction.flat_correct!

CCDReduction.flat_correct!

CCDReduction.subtract_bias

CCDReduction.subtract_bias

CCDReduction.subtract_bias!

CCDReduction.subtract_bias!

CCDReduction.subtract_dark

CCDReduction.subtract_dark

CCDReduction.subtract_dark!

CCDReduction.subtract_dark!

CCDReduction.subtract_overscan

CCDReduction.subtract_overscan

CCDReduction.subtract_overscan!

CCDReduction.trim

CCDReduction.trim

CCDReduction.trimview

API/Reference

CCDReduction.combine — Method

combine(frames...; method = median)
combine(frames; method = median)

Combine multiple frames using method. Multiple frames can also be passed in a vector or as generators for combining.

To pass a custom method, it must have a signature like method(::AbstractArray; dims).

Examples

julia> frame = [reshape(1.0:4.0, (2, 2)) for i = 1:4];

julia> combine(frame)
2×2 Array{Float64,2}:
 1.0  3.0
 2.0  4.0

julia> combine(frame, method = sum)
2×2 Array{Float64,2}:
 4.0  12.0
 8.0  16.0

CCDReduction.combine — Method

combine(frames...; hdu=1, kwargs...)

Load multiple FITS files or HDUs before combining. If loading from filenames, the HDU number can be specified with hdu as either an integer or a tuple corresponding to each file.

Load multiple FITS files or HDUs before combining. Filenames or HDUs can also be passed as vectors for loading before combining.

CCDReduction.crop — Function

crop(frame::FITSIO.ImageHDU, shape; force_equal = true)
crop(filename::String, shape; hdu=1, force_equal = true)

Load a FITS file or HDU before cropping.

CCDReduction.crop — Method

crop(frame::AbstractArray, shape; force_equal = true)

Crops frame to the size specified by shape anchored by the frame center.

This will remove rows/cols of the frame equally on each side. When there is an uneven difference in sizes (e.g. size 9 -> 6 can't be removed equally) the default is to increase the output size (e.g. 6 -> 7) so there is equal removal on each side. To disable this, set force_equal=false, which will remove the extra slice from the end of the axis.

Examples

julia> frame = reshape(1:25, (5, 5));

julia> crop(frame, (3, 3))
3×3 Array{Int64,2}:
 7  12  17
 8  13  18
 9  14  19

julia> crop(frame, (4, 3), force_equal = false)
4×3 Array{Int64,2}:
 6  11  16
 7  12  17
 8  13  18
 9  14  19

See Also

cropview

CCDReduction.cropview — Method

cropview(frame::AbstractArray, shape; force_equal = true)

Crops frame to the size specified by shape anchored by the frame center.

This function is same as the crop function but returns a view of the frame.

Note

This function returns a view of the frame, so any modification to output array will result in modification of frame.

See Also

crop

CCDReduction.flat_correct — Function

flat_correct(frame, flat_frame; [hdu=1], kwargs...)

Load frame and/or flat_frame from a given filename or FITSIO.ImageHDU. If loading from a file, you can specify the appropriate header using the hdu keyword, which can be given as an integer or a tuple (if multiple files are being loaded). All keyword arguments will be passed to the flat_correct! function after loading the data.

CCDReduction.flat_correct! — Function

flat_correct!(frame::AbstractArray, flat_frame::FITSIO.ImageHDU; kwargs...)
flat_correct!(frame::AbstractArray, flat_frame::String; hdu = 1, kwargs...)

Load a FITS file or HDU for the flat frame before correcting frame in-place.

CCDReduction.flat_correct! — Method

flat_correct!(frame::AbstractArray, flat_frame::AbstractArray; norm_value = mean(flat_frame))

In-place version of flat_correct

See Also

flat_correct

CCDReduction.flat_correct — Method

flat_correct(frame::AbstractArray, flat_frame::AbstractArray; norm_value = mean(flat_frame))

Correct frame for non-uniformity using the calibrated flat_frame.

By default, the flat_frame is normalized by its mean, but this can be changed by providing a custom norm_value.

Note

This function may introduce non-finite values if flat_frame contains values very close to 0 due to dividing by zero. The default behavior will return Inf if the frame value is non-zero, and Nan if the frame value is 0.

Examples

julia> frame = ones(3, 3);

julia> flat = fill(2.0, (3, 3));

julia> flat_correct(frame, flat, norm_value = 1.0)
3×3 Array{Float64,2}:
 0.5  0.5  0.5
 0.5  0.5  0.5
 0.5  0.5  0.5

julia> flat_correct(frame, flat)
3×3 Array{Float64,2}:
 1.0  1.0  1.0
 1.0  1.0  1.0
 1.0  1.0  1.0

See Also

flat_correct!

CCDReduction.subtract_bias — Function

subtract_bias(frame, bias_frame; [hdu = 1])

Load frame and/or bias_frame from a given filename or FITSIO.ImageHDU. If loading from a file, you can specify the appropriate header using the hdu keyword, which can be given as an integer or a tuple (if multiple files are being loaded). All keyword arguments will be passed to the subtract_bias! function after loading the data.

CCDReduction.subtract_bias! — Function

subtract_bias!(frame::AbstractArray, bias_frame::FITSIO.ImageHDU)
subtract_bias!(frame::AbstractArray, bias_frame::String; hdu = 1)

Load a FITS file or HDU for the bias frame before subtracting from frame in-place.

CCDReduction.subtract_bias! — Method

subtract_bias!(frame::AbstractArray, bias_frame::AbstractArray)

In-place version of subtract_bias

See Also

subtract_bias

CCDReduction.subtract_bias — Method

subtract_bias(frame::AbstractArray, bias_frame::AbstractArray)

Subtract the bias_frame from frame.

Examples

julia> frame = [1.0 2.2 3.3 4.5];

julia> bias = [0.0 0.2 0.3 0.5];

julia> subtract_bias(frame, bias)
1×4 Array{Float64,2}:
 1.0  2.0  3.0  4.0

See Also

subtract_bias!

CCDReduction.subtract_dark! — Method

subtract_dark!(frame::AbstractArray, dark_frame::AbstractArray; data_exposure = 1, dark_exposure = 1)

In-place version of subtract_dark

See Also

subtract_dark

CCDReduction.subtract_dark! — Method

subtract_dark!(frame::AbstractArray, dark_frame::FITSIO.ImageHDU; data_exposure = 1, dark_exposure = 1)
subtract_dark!(frame::AbstractArray, dark_frame::String; hdu = 1, data_exposure = 1, dark_exposure = 1)

Load a FITS file or HDU for the dark frame before subtracting from frame in-place. If dark_exposure is a symbol it will be parsed from the FITS header (case sensitive).

CCDReduction.subtract_dark — Method

subtract_dark(frame, dark_frame; [hdu = 1], data_exposure = 1, dark_exposure = 1)

Subtract the dark frame from frame. If either arguments are FITSIO.ImageHDU they will be loaded into memory. If either arguments are strings we will attempt to locate a FITS file and open it before loading the data from the given hdu. If loading multiple files, you can specify the HDU numbers separately (hdu=(1, 2)) or simultanesously (hdu=1). If data_exposure or dark_exposure is a symbol it will be read from the FITS header with that key (case sensitive).

CCDReduction.subtract_dark — Method

subtract_dark(frame::AbstractArray, dark_frame::AbstractArray; data_exposure = 1, dark_exposure = 1)

Subtract the dark_frame from frame.

Examples

julia> frame = ones(3, 3);

julia> dark_frame = ones(3, 3);

julia> subtract_dark(frame, dark_frame)
3×3 Array{Float64,2}:
 0.0  0.0  0.0
 0.0  0.0  0.0
 0.0  0.0  0.0

julia> subtract_dark(frame, dark_frame, data_exposure = 1, dark_exposure = 4)
3×3 Array{Float64,2}:
 0.75  0.75  0.75
 0.75  0.75  0.75
 0.75  0.75  0.75

See Also

subtract_dark!

CCDReduction.subtract_overscan — Function

subtract_overscan(frame::FITSIO.ImageHDU, idxs; [dims])
subtract_overscan(filename::String, idxs; hdu=1, [dims])

Load a FITS file or HDU before subtracting the overscan region. If idxs is a symbol it will be read from the FITS header with that key (case sensitive).

CCDReduction.subtract_overscan! — Method

subtract_overscan!(frame::AbstractArray, idxs; dims = axes_min_length(idxs))

In-place version of subtract_overscan

See Also

subtract_overscan

CCDReduction.subtract_overscan — Method

subtract_overscan(frame, idxs; dims = axes_min_length(idxs))

Subtract the overscan frame from image.

dims is the dimension along which overscan_frame is combined. The default value of dims is the axis with smaller length in overscan region. If idxs is a string it will be parsed as FITS-style indices.

Examples

julia> frame = [4.0 2.0 3.0 1.0 1.0];

julia> subtract_overscan(frame, (:, 4:5), dims = 2)
1×5 Array{Float64,2}:
 3.0  1.0  2.0  0.0  0.0

julia> subtract_overscan(frame, "[4:5, 1:1]", dims = 2)
1×5 Array{Float64,2}:
 3.0  1.0  2.0  0.0  0.0

See Also

subtract_overscan!

CCDReduction.trim — Function

trim(frame::FITSIO.ImageHDU, idxs)
trim(filename::String, idxs; hdu=1)

Load a FITS file or HDU before trimming. If idxs is a symbol it will be read from the FITS header with that key (case sensitive).

CCDReduction.trim — Method

trim(frame, idxs)

Trims the frame to remove the region specified by idxs.

This function trims the array in a manner such that final array should be rectangular. The indices follow standard Julia convention, so (:, 45:60) trims all columns from 45 to 60 and (1:20, :) trims all the rows from 1 to 20. The function also supports FITS-style indices.

Examples

julia> frame = ones(5, 5);

julia> trim(frame, (:, 2:5))
5×1 Array{Float64,2}:
 1.0
 1.0
 1.0
 1.0
 1.0

julia> trim(frame, "[2:5, 1:5]")
5×1 Array{Float64,2}:
 1.0
 1.0
 1.0
 1.0
 1.0

See Also

trimview

CCDReduction.trimview — Method

trimview(frame::AbstractArray, idxs)

Trims the frame to remove the region specified by idxs.

This function is same as the trim function but returns a view of the frame.

Note

This function returns a view of the frame, so any modification to output array will result in modification of frame.

See Also

trim